[.ssf {spatstat.explore}	R Documentation

Subset of spatially sampled function

Description

Extract a subset of the data for a spatially sampled function.

Usage

 ## S3 method for class 'ssf'
x[i, j, ..., drop]

Arguments

x
Object of class "ssf".

i
Subset index applying to the locations where the function is sampled.

j
Subset index applying to the columns (variables) measured at each location.

..., drop
Ignored.

Details

This is the subset operator for the class "ssf".

Value

Another object of class "ssf".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

ssf, with.ssf

Examples

  f <- ssf(cells, data.frame(d=nndist(cells), i=1:42))
  f
  f[1:10,]
  f[ ,1]

[Package spatstat.explore version 3.3-4 Index]