FmultiInhom {spatstat.explore}	R Documentation

Inhomogeneous Marked F-Function

Description

For a marked point pattern, estimate the inhomogeneous version of the multitype F function, effectively the cumulative distribution function of the distance from a fixed point to the nearest point in subset J, adjusted for spatially varying intensity.

Usage

  Fmulti.inhom(X, J,
              lambda = NULL, lambdaJ = NULL, lambdamin = NULL,
              ...,
              r = NULL)

  FmultiInhom(X, J,
              lambda = NULL, lambdaJ = NULL, lambdamin = NULL,
              ...,
              r = NULL)

Arguments

X
A spatial point pattern (object of class "ppp".

J
A subset index specifying the subset of points to which distances are measured. Any kind of subset index acceptable to [.ppp.

lambda
Intensity estimates for each point of X. A numeric vector of length equal to npoints(X). Incompatible with lambdaJ.

lambdaJ
Intensity estimates for each point of X[J]. A numeric vector of length equal to npoints(X[J]). Incompatible with lambda.

lambdamin
A lower bound for the intensity, or at least a lower bound for the values in lambdaJ or lambda[J].

...
Extra arguments passed to as.mask to control the pixel resolution for the computation.

r
Vector of distance values at which the inhomogeneous G function should be estimated. There is a sensible default.

Details

See Cronie and Van Lieshout (2015).

The functions FmultiInhom and Fmulti.inhom are identical.

Value

Object of class "fv" containing the estimate of the inhomogeneous multitype F function.

Author(s)

Ottmar Cronie and Marie-Colette van Lieshout. Rewritten for spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Cronie, O. and Van Lieshout, M.N.M. (2015) Summary statistics for inhomogeneous marked point processes. Annals of the Institute of Statistical Mathematics DOI: 10.1007/s10463-015-0515-z

See Also

Finhom

Examples

  X <- amacrine
  J <- (marks(X) == "off")
  online <- interactive()
  eps <- if(online) NULL else 0.025
  if(online && require(spatstat.model)) {
    mod <- ppm(X ~ marks * x, eps=eps)
    lambdaX <- fitted(mod, dataonly=TRUE)
    lambdaOff <- predict(mod, eps=eps)[["off"]]
    lmin <- min(lambdaOff) * 0.9
  } else {
    ## faster computation for package checker only
    lambdaX <- intensity(X)[as.integer(marks(X))]
    lmin <- intensity(X)[2] * 0.9
  }

  plot(FmultiInhom(X, J, lambda=lambdaX, lambdamin=lmin, eps=eps))

[Package spatstat.explore version 3.3-4 Index]