G3est {spatstat.explore}	R Documentation

Nearest Neighbour Distance Distribution Function of a Three-Dimensional Point Pattern

Description

Estimates the nearest-neighbour distance distribution function G3(r) from a three-dimensional point pattern.

Usage

G3est(X, ..., rmax = NULL, nrval = 128, correction = c("rs", "km", "Hanisch"))

Arguments

X
Three-dimensional point pattern (object of class "pp3").

...
Ignored.

rmax
Optional. Maximum value of argument r for which G3(r) will be estimated.

nrval
Optional. Number of values of r for which G3(r) will be estimated. A large value of nrval is required to avoid discretisation effects.

correction
Optional. Character vector specifying the edge correction(s) to be applied. See Details.

Details

For a stationary point process Phi in three-dimensional space, the nearest-neighbour function is

G3(r) = P(d*(x,Phi) <= r | x in Phi)

the cumulative distribution function of the distance d*(x,Phi) from a typical point x in Phi to its nearest neighbour, i.e. to the nearest other point of Phi.

The three-dimensional point pattern X is assumed to be a partial realisation of a stationary point process Phi. The nearest neighbour function of Phi can then be estimated using techniques described in the References. For each data point, the distance to the nearest neighbour is computed. The empirical cumulative distribution function of these values, with appropriate edge corrections, is the estimate of G3(r).

The available edge corrections are:

"rs":

the reduced sample (aka minus sampling, border correction) estimator (Baddeley et al, 1993)

"km":

the three-dimensional version of the Kaplan-Meier estimator (Baddeley and Gill, 1997)

"Hanisch":

the three-dimensional generalisation of the Hanisch estimator (Hanisch, 1984).

Alternatively correction="all" selects all options.

Value

A function value table (object of class "fv") that can be plotted, printed or coerced to a data frame containing the function values.

Warnings

A large value of nrval is required in order to avoid discretisation effects (due to the use of histograms in the calculation).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rana Moyeed.

References

Baddeley, A.J, Moyeed, R.A., Howard, C.V. and Boyde, A. (1993) Analysis of a three-dimensional point pattern with replication. Applied Statistics 42, 641–668.

Baddeley, A.J. and Gill, R.D. (1997) Kaplan-Meier estimators of interpoint distance distributions for spatial point processes. Annals of Statistics 25, 263–292.

Hanisch, K.-H. (1984) Some remarks on estimators of the distribution function of nearest neighbour distance in stationary spatial point patterns. Mathematische Operationsforschung und Statistik, series Statistics 15, 409–412.

See Also

pp3 to create a three-dimensional point pattern (object of class "pp3").

F3est, K3est, pcf3est for other summary functions of a three-dimensional point pattern.

Gest to estimate the empty space function of point patterns in two dimensions.

Examples

  X <- rpoispp3(42)
  Z <- G3est(X)
  if(interactive()) plot(Z)

[Package spatstat.explore version 3.3-4 Index]