GmultiInhom {spatstat.explore}	R Documentation

Inhomogeneous Marked G-Function

Description

For a marked point pattern, estimate the inhomogeneous version of the multitype G function, effectively the cumulative distribution function of the distance from a point in subset I to the nearest point in subset J, adjusted for spatially varying intensity.

Usage

  Gmulti.inhom(X, I, J,
             lambda = NULL, lambdaI = NULL, lambdaJ = NULL,
             lambdamin = NULL, ...,
             r = NULL,
             ReferenceMeasureMarkSetI = NULL,
             ratio = FALSE)

  GmultiInhom(X, I, J,
             lambda = NULL, lambdaI = NULL, lambdaJ = NULL,
             lambdamin = NULL, ...,
             r = NULL,
             ReferenceMeasureMarkSetI = NULL,
             ratio = FALSE)

Arguments

X
A spatial point pattern (object of class "ppp".

I
A subset index specifying the subset of points from which distances are measured. Any kind of subset index acceptable to [.ppp.

J
A subset index specifying the subset of points to which distances are measured. Any kind of subset index acceptable to [.ppp.

lambda
Intensity estimates for each point of X. A numeric vector of length equal to npoints(X). Incompatible with lambdaI,lambdaJ.

lambdaI
Intensity estimates for each point of X[I]. A numeric vector of length equal to npoints(X[I]). Incompatible with lambda.

lambdaJ
Intensity estimates for each point of X[J]. A numeric vector of length equal to npoints(X[J]). Incompatible with lambda.

lambdamin
A lower bound for the intensity, or at least a lower bound for the values in lambdaJ or lambda[J].

...
Ignored.

r
Vector of distance values at which the inhomogeneous G function should be estimated. There is a sensible default.

ReferenceMeasureMarkSetI
Optional. The total measure of the mark set. A positive number.

ratio
Logical value indicating whether to save ratio information.

Details

See Cronie and Van Lieshout (2015).

The functions GmultiInhom and Gmulti.inhom are identical.

Value

Object of class "fv" containing the estimate of the inhomogeneous multitype G function.

Author(s)

Ottmar Cronie and Marie-Colette van Lieshout. Rewritten for spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Cronie, O. and Van Lieshout, M.N.M. (2015) Summary statistics for inhomogeneous marked point processes. Annals of the Institute of Statistical Mathematics DOI: 10.1007/s10463-015-0515-z

See Also

Ginhom, Gmulti

Examples

  X <- rescale(amacrine)
  I <- (marks(X) == "on")
  J <- (marks(X) == "off")
  if(interactive() && require(spatstat.model)) {
    ## how to do it normally
    mod <- ppm(X ~ marks * x)
    lam <- fitted(mod, dataonly=TRUE)
    lmin <- min(predict(mod)[["off"]]) * 0.9
  } else {
    ## for package testing
    lam <- intensity(X)[as.integer(marks(X))]
    lmin <- intensity(X)[2] * 0.9
  }
  plot(GmultiInhom(X, I, J, lambda=lam, lambdamin=lmin))
  # equivalent
  plot(GmultiInhom(X, I, J, lambdaI=lam[I], lambdaJ=lam[J], lambdamin=lmin),
       main="")

[Package spatstat.explore version 3.3-4 Index]