Jdot.inhom {spatstat.explore}	R Documentation

Inhomogeneous Multitype J function (i-to-any)

Description

For a multitype point pattern, estimate the inhomogeneous multitype J function summarising the interpoint dependence between points of type i and points of any type.

Usage

  Jdot.inhom(X, i, 
             lambdaI = NULL, lambdadot = NULL,
             lambdamin = NULL,
             ...,
             r = NULL, ReferenceMeasureMarkSetI = NULL, ratio = FALSE)

Arguments

X
The observed point pattern, from which an estimate of the inhomogeneous multitype J function Ji.(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i
The type (mark value) of the points in X from which distances are measured. A character string (or something that will be converted to a character string). Defaults to the first level of marks(X).

lambdaI
Optional. Values of the estimated intensity of the sub-process of points of type i. Either a pixel image (object of class "im"), a numeric vector containing the intensity values at each of the type i points in X, a fitted point process model (object of class "ppm" or "kppm" or "dppm"), or a function(x,y) which can be evaluated to give the intensity value at any location.

lambdadot
Optional. Values of the estimated intensity of the point process. Either a pixel image (object of class "im"), a numeric vector containing the intensity values at each of the points in X, a fitted point process model (object of class "ppm" or "kppm" or "dppm"), or a function(x,y) which can be evaluated to give the intensity value at any location.

lambdamin
Optional. The minimum possible value of the intensity over the spatial domain. A positive numerical value.

...
Extra arguments passed to as.mask to control the pixel resolution for the computation.

r
vector of values for the argument r at which the inhomogeneous K function should be evaluated. Not normally given by the user; there is a sensible default.

ReferenceMeasureMarkSetI
Optional. The total measure of the mark set. A positive number.

ratio
Logical value indicating whether to save ratio information.

Details

This function is the counterpart of Jdot for inhomogeneous patterns. It is computed as a special case of Jmulti.inhom.

Value

Object of class "fv" containing the estimate of the inhomogeneous multitype J function.

Author(s)

Jonatan Gonzalez and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Cronie, O. and Van Lieshout, M.N.M. (2015) Summary statistics for inhomogeneous marked point processes. Annals of the Institute of Statistical Mathematics DOI: 10.1007/s10463-015-0515-z

See Also

Jdot.inhom, Jmulti.inhom, Jdot.

Examples

  X <- rescale(amacrine)
  if(interactive() && require(spatstat.model)) {
    ## how to do it normally
    mod <- ppm(X ~ marks * x)
    lam <- fitted(mod, dataonly=TRUE)
    lmin <- min(predict(mod)[["off"]]) * 0.9
    dd <- NULL
  } else {
    ## for package testing 
    lam <- intensity(X)[as.integer(marks(X))]
    lmin <- intensity(X)[2] * 0.9
    dd <- 32
  }
  lamI <- lam[marks(X) == "on"]
  JD <- Jdot.inhom(X, "on", lambdaI=lamI, lambdadot=lam, lambdamin=lmin,
                   dimyx=dd)

[Package spatstat.explore version 3.3-4 Index]