Jmulti.inhom {spatstat.explore}	R Documentation

Inhomogeneous Marked J-Function

Description

For a marked point pattern, estimate the inhomogeneous version of the multitype J function.

Usage

Jmulti.inhom(X, I, J,
             lambda = NULL, lambdaI = NULL, lambdaJ = NULL,
             lambdamin = NULL,
             ...,
             r = NULL,
             ReferenceMeasureMarkSetI = NULL,
             ratio = FALSE)

Arguments

X
The observed point pattern, from which an estimate of the inhomogeneous multitype J function J[IJ](r) will be computed. It must be a marked point pattern. See under Details.

I
Subset index specifying the points of X from which distances are measured, for the inhomogeneous G function.

J
Subset index specifying the points in X to which distances are measured, for the inhomogeneous G and F functions.

lambda
Optional. Values of the estimated intensity function. Either a vector giving the intensity values at the points of the pattern X, a pixel image (object of class "im") giving the intensity values at all locations, a fitted point process model (object of class "ppm") or a function(x,y) which can be evaluated to give the intensity value at any location.

lambdaI
Optional. Values of the estimated intensity of the sub-process X[I]. Either a pixel image (object of class "im"), a numeric vector containing the intensity values at each of the points in X[I], a fitted point process model (object of class "ppm" or "kppm" or "dppm"), or a function(x,y) which can be evaluated to give the intensity value at any location,

lambdaJ
Optional. Values of the estimated intensity of the sub-process X[J]. Either a pixel image (object of class "im"), a numeric vector containing the intensity values at each of the points in X[J], a fitted point process model (object of class "ppm" or "kppm" or "dppm"), or a function(x,y) which can be evaluated to give the intensity value at any location.

lambdamin
Optional. The minimum possible value of the intensity over the spatial domain. A positive numerical value.

...
Extra arguments passed to as.mask to control the pixel resolution for the computation.

r
vector of values for the argument r at which the inhomogeneous K function should be evaluated. Not normally given by the user; there is a sensible default.

ReferenceMeasureMarkSetI
Optional. The total measure of the mark set. A positive number.

ratio
Logical value indicating whether to save ratio information.

Details

This function is the counterpart of Jmulti for inhomogeneous patterns. It is computed by evaluating the inhomogeneous G function GmultiInhom and the inhomogeneous F function FmultiInhom and computing the ratio J = (1-G)/(1-F).

Value

Object of class "fv" containing the estimate of the inhomogeneous multitype J function.

Author(s)

Jonatan Gonzalez and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Cronie, O. and Van Lieshout, M.N.M. (2015) Summary statistics for inhomogeneous marked point processes. Annals of the Institute of Statistical Mathematics DOI: 10.1007/s10463-015-0515-z

See Also

Jcross.inhom, Jdot.inhom for special cases.

GmultiInhom, FmultiInhom, Jmulti.

Examples

  X <- rescale(amacrine)
  I <- (marks(X) == "on")
  J <- (marks(X) == "off")
  if(interactive() && require(spatstat.model)) {
    ## how to do it normally
    mod <- ppm(X ~ marks * x)
    lam <- fitted(mod, dataonly=TRUE)
    lmin <- min(predict(mod)[["off"]]) * 0.9
    dd <- NULL
  } else {
    ## for package testing
    lam <- intensity(X)[as.integer(marks(X))]
    lmin <- intensity(X)[2] * 0.9
    dd <- 32
  }
  JM <- Jmulti.inhom(X, I, J, lambda=lam, lambdamin=lmin, dimyx=dd)

[Package spatstat.explore version 3.3-4 Index]