Kdot {spatstat.explore}	R Documentation

Multitype K Function (i-to-any)

Description

For a multitype point pattern, estimate the multitype K function which counts the expected number of other points of the process within a given distance of a point of type i.

Usage

Kdot(X, i, r=NULL, breaks=NULL, correction, ..., ratio=FALSE, from)

Arguments

X
The observed point pattern, from which an estimate of the multitype K function Ki.(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i
The type (mark value) of the points in X from which distances are measured. A character string (or something that will be converted to a character string). Defaults to the first level of marks(X).

r
numeric vector. The values of the argument r at which the distribution function Ki.(r) should be evaluated. There is a sensible default. First-time users are strongly advised not to specify this argument. See below for important conditions on r.

breaks
This argument is for internal use only.

correction
A character vector containing any selection of the options "border", "bord.modif", "isotropic", "Ripley", "translate", "translation", "periodic", "none" or "best". It specifies the edge correction(s) to be applied. Alternatively correction="all" selects all options.

...
Ignored.

ratio
Logical. If TRUE, the numerator and denominator of each edge-corrected estimate will also be saved, for use in analysing replicated point patterns.

from
An alternative way to specify i.

Details

This function Kdot and its companions Kcross and Kmulti are generalisations of the function Kest to multitype point patterns.

A multitype point pattern is a spatial pattern of points classified into a finite number of possible “colours” or “types”. In the spatstat package, a multitype pattern is represented as a single point pattern object in which the points carry marks, and the mark value attached to each point determines the type of that point.

The argument X must be a point pattern (object of class "ppp") or any data that are acceptable to as.ppp. It must be a marked point pattern, and the mark vector X$marks must be a factor.

The argument i will be interpreted as a level of the factor X$marks. If i is missing, it defaults to the first level of the marks factor, i = levels(X$marks)[1].

The “type i to any type” multitype K function of a stationary multitype point process X is defined so that lambda Ki.(r) equals the expected number of additional random points within a distance r of a typical point of type i in the process X. Here lambda is the intensity of the process, i.e. the expected number of points of X per unit area. The function Ki. is determined by the second order moment properties of X.

An estimate of Ki.(r) is a useful summary statistic in exploratory data analysis of a multitype point pattern. If the subprocess of type i points were independent of the subprocess of points of all types not equal to i, then Ki.(r) would equal pi * r^2. Deviations between the empirical Ki. curve and the theoretical curve pi * r^2 may suggest dependence between types.

This algorithm estimates the distribution function Ki.(r) from the point pattern X. It assumes that X can be treated as a realisation of a stationary (spatially homogeneous) random spatial point process in the plane, observed through a bounded window. The window (which is specified in X as Window(X)) may have arbitrary shape. Biases due to edge effects are treated in the same manner as in Kest, using the chosen edge correction(s).

The argument r is the vector of values for the distance r at which Ki.(r) should be evaluated. The values of r must be increasing nonnegative numbers and the maximum r value must not exceed the radius of the largest disc contained in the window.

The pair correlation function can also be applied to the result of Kdot; see pcf.

Value

An object of class "fv" (see fv.object).

Essentially a data frame containing numeric columns

r
the values of the argument r at which the function Ki.(r) has been estimated

theo
the theoretical value of Ki.(r) for a marked Poisson process, namely pi * r^2

together with a column or columns named "border", "bord.modif", "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function Ki.(r) obtained by the edge corrections named.

If ratio=TRUE then the return value also has two attributes called "numerator" and "denominator" which are "fv" objects containing the numerators and denominators of each estimate of K(r).

Warnings

The argument i is interpreted as a level of the factor X$marks. It is converted to a character string if it is not already a character string. The value i=1 does not refer to the first level of the factor.

The reduced sample estimator of Ki. is pointwise approximately unbiased, but need not be a valid distribution function; it may not be a nondecreasing function of r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

References

Cressie, N.A.C. Statistics for spatial data. John Wiley and Sons, 1991.

Diggle, P.J. Statistical analysis of spatial point patterns. Academic Press, 1983.

Harkness, R.D and Isham, V. (1983) A bivariate spatial point pattern of ants' nests. Applied Statistics 32, 293–303

Lotwick, H. W. and Silverman, B. W. (1982). Methods for analysing spatial processes of several types of points. J. Royal Statist. Soc. Ser. B 44, 406–413.

Ripley, B.D. Statistical inference for spatial processes. Cambridge University Press, 1988.

Stoyan, D, Kendall, W.S. and Mecke, J. Stochastic geometry and its applications. 2nd edition. Springer Verlag, 1995.

See Also

Kdot, Kest, Kmulti, pcf

Examples

     # Lansing woods data: 6 types of trees
    woods <- lansing

    

    Kh. <- Kdot(woods, "hickory") 
    # diagnostic plot for independence between hickories and other trees
    plot(Kh.)

    # synthetic example with two marks "a" and "b"
    
      pp <- runifpoispp(50)
      pp <- pp %mark% factor(sample(c("a","b"), npoints(pp), replace=TRUE))
      K <- Kdot(pp, "a")
   

[Package spatstat.explore version 3.3-4 Index]