Kest.fft {spatstat.explore}	R Documentation

K-function using FFT

Description

Estimates the reduced second moment function K(r) from a point pattern in a window of arbitrary shape, using the Fast Fourier Transform.

Usage

  Kest.fft(X, sigma, r=NULL, ..., breaks=NULL)

Arguments

X
The observed point pattern, from which an estimate of K(r) will be computed. An object of class "ppp", or data in any format acceptable to as.ppp().

sigma
Standard deviation of the isotropic Gaussian smoothing kernel.

r
Optional. Vector of values for the argument r at which K(r) should be evaluated. There is a sensible default.

...
Arguments passed to as.mask determining the spatial resolution for the FFT calculation.

breaks
This argument is for internal use only.

Details

This is an alternative to the function Kest for estimating the K function. It may be useful for very large patterns of points.

Whereas Kest computes the distance between each pair of points analytically, this function discretises the point pattern onto a rectangular pixel raster and applies Fast Fourier Transform techniques to estimate K(t). The hard work is done by the function Kmeasure.

The result is an approximation whose accuracy depends on the resolution of the pixel raster. The resolution is controlled by the arguments ..., or by setting the parameter npixel in spatstat.options.

Value

An object of class "fv" (see fv.object).

Essentially a data frame containing columns

r
the vector of values of the argument r at which the function K has been estimated

border
the estimates of K(r) for these values of r

theo
the theoretical value K(r) = pi * r^2 for a stationary Poisson process

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Cressie, N.A.C. Statistics for spatial data. John Wiley and Sons, 1991.

Diggle, P.J. Statistical analysis of spatial point patterns. Academic Press, 1983.

Ohser, J. (1983) On estimators for the reduced second moment measure of point processes. Mathematische Operationsforschung und Statistik, series Statistics, 14, 63 – 71.

Ripley, B.D. Statistical inference for spatial processes. Cambridge University Press, 1988.

Stoyan, D, Kendall, W.S. and Mecke, J. (1995) Stochastic geometry and its applications. 2nd edition. Springer Verlag.

Stoyan, D. and Stoyan, H. (1994) Fractals, random shapes and point fields: methods of geometrical statistics. John Wiley and Sons.

See Also

Kest, Kmeasure, spatstat.options

Examples

 pp <- runifpoint(10000)
 
 Kpp <- Kest.fft(pp, 0.01)
 plot(Kpp)
 

[Package spatstat.explore version 3.3-4 Index]