Lcross {spatstat.explore}	R Documentation

Multitype L-function (cross-type)

Description

Calculates an estimate of the cross-type L-function for a multitype point pattern.

Usage

  Lcross(X, i, j, ..., from, to, correction)

Arguments

X
The observed point pattern, from which an estimate of the cross-type L function Lij(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i
The type (mark value) of the points in X from which distances are measured. A character string (or something that will be converted to a character string). Defaults to the first level of marks(X).

j
The type (mark value) of the points in X to which distances are measured. A character string (or something that will be converted to a character string). Defaults to the second level of marks(X).

correction,...
Arguments passed to Kcross.

from,to
An alternative way to specify i and j respectively.

Details

The cross-type L-function is a transformation of the cross-type K-function,

Lij(r) = sqrt(Kij(r)/pi)

where Kij(r) is the cross-type K-function from type i to type j. See Kcross for information about the cross-type K-function.

The command Lcross first calls Kcross to compute the estimate of the cross-type K-function, and then applies the square root transformation.

For a marked point pattern in which the points of type i are independent of the points of type j, the theoretical value of the L-function is Lij(r) = r. The square root also has the effect of stabilising the variance of the estimator, so that Lij is more appropriate for use in simulation envelopes and hypothesis tests.

Value

An object of class "fv", see fv.object, which can be plotted directly using plot.fv.

Essentially a data frame containing columns

r
the vector of values of the argument r at which the function Lij has been estimated

theo
the theoretical value Lij(r) = r for a stationary Poisson process

together with columns named "border", "bord.modif", "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function Lij obtained by the edge corrections named.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

Kcross, Ldot, Lest

Examples

 L <- Lcross(amacrine, "off", "on")
 plot(L)

[Package spatstat.explore version 3.3-4 Index]