Ldot {spatstat.explore}	R Documentation

Multitype L-function (i-to-any)

Description

Calculates an estimate of the multitype L-function (from type i to any type) for a multitype point pattern.

Usage

  Ldot(X, i, ..., from, correction)

Arguments

X
The observed point pattern, from which an estimate of the dot-type L function Lij(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i
The type (mark value) of the points in X from which distances are measured. A character string (or something that will be converted to a character string). Defaults to the first level of marks(X).

correction,...
Arguments passed to Kdot.

from
An alternative way to specify i.

Details

This command computes

Li.(r) = sqrt(Ki.(r)/pi)

where Ki.(r) is the multitype K-function from points of type i to points of any type. See Kdot for information about Ki.(r).

The command Ldot first calls Kdot to compute the estimate of the i-to-any K-function, and then applies the square root transformation.

For a marked Poisson point process, the theoretical value of the L-function is Li.(r) = r. The square root also has the effect of stabilising the variance of the estimator, so that Li. is more appropriate for use in simulation envelopes and hypothesis tests.

Value

An object of class "fv", see fv.object, which can be plotted directly using plot.fv.

Essentially a data frame containing columns

r
the vector of values of the argument r at which the function Li. has been estimated

theo
the theoretical value Li.(r) = r for a stationary Poisson process

together with columns named "border", "bord.modif", "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function Li. obtained by the edge corrections named.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

Kdot, Lcross, Lest

Examples

 L <- Ldot(amacrine, "off")
 plot(L)

[Package spatstat.explore version 3.3-4 Index]