Ldot.inhom {spatstat.explore}	R Documentation

Inhomogeneous Multitype L Dot Function

Description

For a multitype point pattern, estimate the inhomogeneous version of the dot L function.

Usage

Ldot.inhom(X, i, ..., correction)

Arguments

X
The observed point pattern, from which an estimate of the inhomogeneous cross type L function Li.(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i
The type (mark value) of the points in X from which distances are measured. A character string (or something that will be converted to a character string). Defaults to the first level of marks(X).

correction,...
Other arguments passed to Kdot.inhom.

Details

This a generalisation of the function Ldot to include an adjustment for spatially inhomogeneous intensity, in a manner similar to the function Linhom.

All the arguments are passed to Kdot.inhom, which estimates the inhomogeneous multitype K function Ki.(r) for the point pattern. The resulting values are then transformed by taking L(r) = sqrt(K(r)/pi).

Value

An object of class "fv" (see fv.object).

Essentially a data frame containing numeric columns

r
the values of the argument r at which the function Li.(r) has been estimated

theo
the theoretical value of Li.(r) for a marked Poisson process, identical to r.

together with a column or columns named "border", "bord.modif", "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function Li.(r) obtained by the edge corrections named.

Warnings

The argument i is interpreted as a level of the factor X$marks. It is converted to a character string if it is not already a character string. The value i=1 does not refer to the first level of the factor.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Moller, J. and Waagepetersen, R. Statistical Inference and Simulation for Spatial Point Processes Chapman and Hall/CRC Boca Raton, 2003.

See Also

Ldot, Linhom, Kdot.inhom, Lcross.inhom.

Examples

    # Lansing Woods data
    lan <- lansing
    lan <- lan[seq(1,npoints(lan), by=10)]
    ma <- split(lan)$maple
    lg <- unmark(lan)

    # Estimate intensities by nonparametric smoothing
    lambdaM <- density(ma, sigma=0.15, at="points")
    lambdadot <- density(lg, sigma=0.15, at="points")
    L <- Ldot.inhom(lan, "maple", lambdaI=lambdaM,
                                  lambdadot=lambdadot)

    # synthetic example: type A points have intensity 50,
    #                    type B points have intensity 50 + 100 * x
    lamB <- as.im(function(x,y){50 + 100 * x}, owin())
    lamdot <- as.im(function(x,y) { 100 + 100 * x}, owin())
    X <- superimpose(A=runifpoispp(50), B=rpoispp(lamB))
    L <- Ldot.inhom(X, "B",  lambdaI=lamB,     lambdadot=lamdot)

[Package spatstat.explore version 3.3-4 Index]