Lest {spatstat.explore}	R Documentation

L-function

Description

Calculates an estimate of the L-function (Besag's transformation of Ripley's K-function) for a spatial point pattern.

Usage

  Lest(X, ..., correction)

Arguments

X
The observed point pattern, from which an estimate of L(r) will be computed. An object of class "ppp", or data in any format acceptable to as.ppp().

correction,...
Other arguments passed to Kest to control the estimation procedure.

Details

This command computes an estimate of the L-function for the spatial point pattern X. The L-function is a transformation of Ripley's K-function,

L(r) = sqrt(K(r)/pi)

where K(r) is the K-function.

See Kest for information about Ripley's K-function. The transformation to L was proposed by Besag (1977).

The command Lest first calls Kest to compute the estimate of the K-function, and then applies the square root transformation.

For a completely random (uniform Poisson) point pattern, the theoretical value of the L-function is L(r) = r. The square root also has the effect of stabilising the variance of the estimator, so that L(r) is more appropriate for use in simulation envelopes and hypothesis tests.

See Kest for the list of arguments.

Value

An object of class "fv", see fv.object, which can be plotted directly using plot.fv.

Essentially a data frame containing columns

r
the vector of values of the argument r at which the function L has been estimated

theo
the theoretical value L(r) = r for a stationary Poisson process

together with columns named "border", "bord.modif", "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function L(r) obtained by the edge corrections named.

Variance approximations

If the argument var.approx=TRUE is given, the return value includes columns rip and ls containing approximations to the variance of Lest(r) under CSR. These are obtained by the delta method from the variance approximations described in Kest.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Besag, J. (1977) Discussion of Dr Ripley's paper. Journal of the Royal Statistical Society, Series B, 39, 193–195.

See Also

Kest, pcf

Examples

 L <- Lest(cells)
 plot(L, main="L function for cells")

[Package spatstat.explore version 3.3-4 Index]