Linhom {spatstat.explore}	R Documentation

Inhomogeneous L-function

Description

Calculates an estimate of the inhomogeneous version of the L-function (Besag's transformation of Ripley's K-function) for a spatial point pattern.

Usage

  Linhom(X, ..., correction)

Arguments

X
The observed point pattern, from which an estimate of L(r) will be computed. An object of class "ppp", or data in any format acceptable to as.ppp().

correction,...
Other arguments passed to Kinhom to control the estimation procedure.

Details

This command computes an estimate of the inhomogeneous version of the L-function for a spatial point pattern.

The original L-function is a transformation (proposed by Besag) of Ripley's K-function,

L(r) = sqrt(K(r)/pi)

where K(r) is the Ripley K-function of a spatially homogeneous point pattern, estimated by Kest.

The inhomogeneous L-function is the corresponding transformation of the inhomogeneous K-function, estimated by Kinhom. It is appropriate when the point pattern clearly does not have a homogeneous intensity of points. It was proposed by Baddeley, Moller and Waagepetersen (2000).

The command Linhom first calls Kinhom to compute the estimate of the inhomogeneous K-function, and then applies the square root transformation.

For a Poisson point pattern (homogeneous or inhomogeneous), the theoretical value of the inhomogeneous L-function is L(r) = r. The square root also has the effect of stabilising the variance of the estimator, so that L is more appropriate for use in simulation envelopes and hypothesis tests.

Value

An object of class "fv", see fv.object, which can be plotted directly using plot.fv.

Essentially a data frame containing columns

r
the vector of values of the argument r at which the function L has been estimated

theo
the theoretical value L(r) = r for a stationary Poisson process

together with columns named "border", "bord.modif", "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function L(r) obtained by the edge corrections named.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Baddeley, A., Moller, J. and Waagepetersen, R. (2000) Non- and semiparametric estimation of interaction in inhomogeneous point patterns. Statistica Neerlandica 54, 329–350.

See Also

Kest, Lest, Kinhom, pcf

Examples

 X <- japanesepines
 L <- Linhom(X, sigma=0.1)
 plot(L, main="Inhomogeneous L function for Japanese Pines")

[Package spatstat.explore version 3.3-4 Index]