Math.fv {spatstat.explore}	R Documentation

S3 Group Generic Methods for Function Tables

Description

These are group generic methods for objects of class "fv", which allows for usual mathematical functions and operators to be applied directly to function tables. See Details for a list of implemented functions.

Usage

## S3 methods for group generics have prototypes:
Math(x, ...)
Ops(e1, e2)
Complex(z)
Summary(..., na.rm=FALSE, drop=TRUE)

Arguments

x, z, e1, e2
objects of class "fv".

...
further arguments passed to methods.

na.rm
Logical value specifying whether missing values should be removed.

Details

Below is a list of mathematical functions and operators which are defined for objects of class "fv". The methods are implemented using eval.fv, which tries to harmonise the functions via harmonise.fv if they aren't compatible to begin with.

Group "Math":

abs, sign, sqrt,
floor, ceiling, trunc,
round, signif

exp, log, expm1, log1p,
cos, sin, tan,
cospi, sinpi, tanpi,
acos, asin, atan

cosh, sinh, tanh,
acosh, asinh, atanh

lgamma, gamma, digamma, trigamma

cumsum, cumprod, cummax, cummin

Group "Ops":

"+", "-", "*", "/", "^", "%%", "%/%"

"&", "|", "!"

"==", "!=", "<", "<=", ">=", ">"

Group "Summary":

all, any

sum, prod

min, max

range

Group "Complex":

Arg, Conj, Im, Mod, Re

For the Ops group, one of the arguments is permitted to be a single atomic value instead of a function table.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

eval.fv for evaluating expressions involving function tables.

Examples

  ## Convert K function to L function
  K <- Kest(cells)
  L <- sqrt(K/pi)
  ## Manually calculate J function
  FR <- Fest(redwood)
  GR <- Gest(redwood)
  suppressWarnings(JR <- (1-GR)/(1-FR))

[Package spatstat.explore version 3.3-4 Index]