Smooth.ssf {spatstat.explore}	R Documentation

Smooth a Spatially Sampled Function

Description

Applies kernel smoothing to a spatially sampled function.

Usage

 ## S3 method for class 'ssf'
Smooth(X, ...)

Arguments

X
Object of class "ssf".

...
Arguments passed to Smooth.ppp to control the smoothing.

Details

An object of class "ssf" represents a real-valued or vector-valued function that has been evaluated or sampled at an irregular set of points.

The function values will be smoothed using a Gaussian kernel.

Value

A pixel image or a list of pixel images.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

ssf, Smooth.ppp

Examples

   f <- ssf(redwood, nndist(redwood))
   Smooth(f, sigma=0.1)

[Package spatstat.explore version 3.3-4 Index]