Smoothfun.ppp {spatstat.explore}	R Documentation

Smooth Interpolation of Marks as a Spatial Function

Description

Perform spatial smoothing of numeric values observed at a set of irregular locations, and return the result as a function of spatial location.

Usage

Smoothfun(X, ...)

## S3 method for class 'ppp'
Smoothfun(X, sigma = NULL, ...,
                        weights = NULL, edge = TRUE, diggle = FALSE)

Arguments

X
Marked point pattern (object of class "ppp").

sigma
Smoothing bandwidth, or bandwidth selection function, passed to Smooth.ppp.

...
Additional arguments passed to Smooth.ppp.

weights
Optional vector of weights associated with the points of X.

edge,diggle
Logical arguments controlling the edge correction. Arguments passed to Smooth.ppp.

Details

The commands Smoothfun and Smooth both perform kernel-smoothed spatial interpolation of numeric values observed at irregular spatial locations. The difference is that Smooth returns a pixel image, containing the interpolated values at a grid of locations, while Smoothfun returns a function(x,y) which can be used to compute the interpolated value at any spatial location. For purposes such as model-fitting it is more accurate to use Smoothfun to interpolate data.

Value

A function with arguments x,y. The function also belongs to the class "Smoothfun" which has methods for print and as.im. It also belongs to the class "funxy" which has methods for plot, contour and persp.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

Smooth

Examples

  f <- Smoothfun(longleaf)
  f
  f(120, 80)
  plot(f)

[Package spatstat.explore version 3.3-4 Index]