SpatialMedian.ppp {spatstat.explore}	R Documentation

Spatially Weighted Median of Values at Points

Description

Given a spatial point pattern with numeric marks, compute a weighted median of the mark values, with spatially-varying weights that depend on distance to the data points.

Usage

 ## S3 method for class 'ppp'
SpatialMedian(X, sigma = NULL, ...,
        type = 4, at = c("pixels", "points"), leaveoneout = TRUE,
        weights = NULL, edge = TRUE, diggle = FALSE, verbose = FALSE)

Arguments

X
A spatial point pattern (object of class "ppp") with numeric marks.

sigma
Smoothing bandwidth, passed to density.ppp.

...
Further arguments passed to density.ppp controlling the spatial smoothing.

type
Integer specifying the type of median (using the convention of quantile.default; see Details). Only types 1 and 4 are currently implemented.

at
Character string indicating whether to compute the median at every pixel of a pixel image (at="pixels", the default) or at every data point of X (at="points").

leaveoneout
Logical value indicating whether to compute a leave-one-out estimator. Applicable only when at="points".

weights
Optional vector of numeric weights attached to the points of X.

edge,diggle
Arguments passed to density.ppp to determine the edge correction.

verbose
Logical value specifying whether to print progress reports during the calculation.

Details

The argument X should be a spatial point pattern (object of class "ppp") with numeric marks.

The algorithm computes the weighted median of the mark values at each desired spatial location, using spatially-varying weights which depend on distance to the data points.

Suppose the data points are at spatial locations x[1], ..., x[n] and have mark values y[1], ..., y[n]. For a query location u, the smoothed median is defined as the weighted median of the mark values y[1], ..., y[n] with weights w[1], ..., w[n], where

w[i] = k(u,x[i])/(k(u, x[1]) + ... + k(u, x[n]))

where k(u,v) is the smoothing kernel with bandwidth sigma

If at="points" and leaveoneout=TRUE, then a leave-one-out calculation is performed, which means that when the query location is a data point x[i], the value at the data point is ignored, and the weighted median is computed from the values y[j] for all j not equal to i.

Value

If X has a single column of marks:

If at="pixels" (the default), the result is a pixel image (object of class "im").

If at="points", the result is a numeric vector of length equal to the number of points in X.

If X has a data frame of marks:

If at="pixels" (the default), the result is a named list of pixel images (object of class "im"). There is one image for each column of marks. This list also belongs to the class "solist", for which there is a plot method.

If at="points", the result is a data frame with one row for each point of X, and one column for each column of marks. Entries are values of the interpolated function at the points of X.

The return value has attributes "sigma" and "varcov" which report the smoothing bandwidth that was used.

The calculation of the median value depends on the argument type which is interpreted in the same way as for quantile.default. Currently, only types 1 and 4 are implemented. If type=1, the median is always one of the mark values (one of the values in marks(x)). If type=4 (the default), the median value is obtained by linearly interpolating between mark values. Note that the default values of type in SpatialMedian.ppp and SpatialQuantile.ppp are different.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

Generic function SpatialMedian.

SpatialQuantile and SpatialQuantile.ppp for other quantiles.

Smooth.ppp for the spatially weighted average.

Examples

   X <- longleaf
   if(!interactive()) {
    ## mark values rounded to nearest multiple of 10 to reduce check time
    marks(X) <- round(marks(X), -1)
   }
   Z <- SpatialMedian(X, sigma=30)
   ZX <- SpatialMedian(X, sigma=30, at="points")

[Package spatstat.explore version 3.3-4 Index]