SpatialQuantile {spatstat.explore}	R Documentation

Spatially Weighted Median or Quantile

Description

Compute a weighted median or weighted quantile of spatial data.

Usage

 SpatialMedian(X, ...)

 SpatialQuantile(X, prob = 0.5, ...)

Arguments

X
A spatial data object.

prob
Probability for which the quantile is required. A single numeric value between 0 and 1. Default is to calculate the median.

...
Further arguments passed to methods.

Details

The functions SpatialMedian and SpatialQuantile are generic. They calculate spatially weighted medians and quantiles of spatial data. The details depend on the class of X.

There are methods for spatial point patterns (class "ppp") and possibly for other objects.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

Methods SpatialMedian.ppp, SpatialQuantile.ppp.

Smooth for the spatially weighted average.

[Package spatstat.explore version 3.3-4 Index]