Tstat {spatstat.explore}	R Documentation

Third order summary statistic

Description

Computes the third order summary statistic T(r) of a spatial point pattern.

Usage

Tstat(X, ..., r = NULL, rmax = NULL,
    correction = c("border", "translate"), ratio = FALSE, verbose=TRUE)

Arguments

X
The observed point pattern, from which an estimate of T(r) will be computed. An object of class "ppp", or data in any format acceptable to as.ppp().

...
Ignored.

r
Optional. Vector of values for the argument r at which T(r) should be evaluated. Users are advised not to specify this argument; there is a sensible default.

rmax
Optional. Numeric. The maximum value of r for which T(r) should be estimated.

correction
Optional. A character vector containing any selection of the options "none", "border", "bord.modif", "translate", "translation", or "best". It specifies the edge correction(s) to be applied. Alternatively correction="all" selects all options.

ratio
Logical. If TRUE, the numerator and denominator of each edge-corrected estimate will also be saved, for use in analysing replicated point patterns.

verbose
Logical. If TRUE, an estimate of the computation time is printed.

Details

This command calculates the third-order summary statistic T(r) for a spatial point patterns, defined by Schladitz and Baddeley (2000).

The definition of T(r) is similar to the definition of Ripley's K function K(r), except that K(r) counts pairs of points while T(r) counts triples of points. Essentially T(r) is a rescaled cumulative distribution function of the diameters of triangles in the point pattern. The diameter of a triangle is the length of its longest side.

Value

An object of class "fv", see fv.object, which can be plotted directly using plot.fv.

Computation time

If the number of points is large, the algorithm can take a very long time to inspect all possible triangles. A rough estimate of the total computation time will be printed at the beginning of the calculation. If this estimate seems very large, stop the calculation using the user interrupt signal, and call Tstat again, using rmax to restrict the range of r values, thus reducing the number of triangles to be inspected.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Schladitz, K. and Baddeley, A. (2000) A third order point process characteristic. Scandinavian Journal of Statistics 27 (2000) 657–671.

See Also

Kest

Examples

  plot(Tstat(redwood))

[Package spatstat.explore version 3.3-4 Index]