adaptive.density {spatstat.explore}	R Documentation

Adaptive Estimate of Intensity of Point Pattern

Description

Computes an adaptive estimate of the intensity function of a point pattern.

Usage

adaptive.density(X, ..., method=c("voronoi","kernel", "nearest"))

Arguments

X
Point pattern (object of class "ppp" or "lpp").

method
Character string specifying the estimation method

...
Additional arguments passed to densityVoronoi, densityAdaptiveKernel.ppp or nndensity.ppp.

Details

This function is an alternative to density.ppp and density.lpp. It computes an estimate of the intensity function of a point pattern dataset. The result is a pixel image giving the estimated intensity.

If method="voronoi" the data are passed to the function densityVoronoi which estimates the intensity using the Voronoi-Dirichlet tessellation.

If method="kernel" the data are passed to the function densityAdaptiveKernel.ppp which estimates the intensity using a variable-bandwidth kernel estimator. (This is not yet supported when X has class "lpp".)

If method="nearest" the data are passed to the function nndensity.ppp which estimates the intensity using the distance to the k-th nearest data point. (This is not yet supported when X has class "lpp".)

Value

A pixel image (object of class "im" or "linim") whose values are estimates of the intensity of X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk and Mehdi Moradi m2.moradi@yahoo.com.

See Also

density.ppp, densityVoronoi, densityAdaptiveKernel.ppp, nndensity.ppp, im.object.

Examples

  plot(adaptive.density(nztrees, 1), main="Voronoi estimate")

[Package spatstat.explore version 3.3-4 Index]