allstats {spatstat.explore}	R Documentation

Calculate four standard summary functions of a point pattern.

Description

Calculates the F, G, J, and K summary functions for an unmarked point pattern. Returns them as a function array (of class "fasp", see fasp.object).

Usage

  allstats(pp, ..., dataname=NULL, verb=FALSE)

Arguments

pp
The observed point pattern, for which summary function estimates are required. An object of class "ppp". It must not be marked.

...
Optional arguments passed to the summary functions Fest, Gest, Jest and Kest.

dataname
A character string giving an optional (alternative) name for the point pattern.

verb
A logical value meaning “verbose”. If TRUE, progress reports are printed during calculation.

Details

This computes four standard summary statistics for a point pattern: the empty space function F(r), nearest neighbour distance distribution function G(r), van Lieshout-Baddeley function J(r) and Ripley's function K(r). The real work is done by Fest, Gest, Jest and Kest respectively. Consult the help files for these functions for further information about the statistical interpretation of F, G, J and K.

If verb is TRUE, then “progress reports” (just indications of completion) are printed out when the calculations are finished for each of the four function types.

The overall title of the array of four functions (for plotting by plot.fasp) will be formed from the argument dataname. If this is not given, it defaults to the expression for pp given in the call to allstats.

Value

A list of length 4 containing the F, G, J and K functions respectively.

The list can be plotted directly using plot (which dispatches to plot.anylist).

Each list entry retains the format of the output of the relevant estimating routine Fest, Gest, Jest or Kest. Thus each entry in the list is a function value table (object of class "fv", see fv.object).

The default formulae for plotting these functions are cbind(km,theo) ~ r for F, G, and J, and cbind(trans,theo) ~ r for K.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

plot.anylist, plot.fv, fv.object, Fest, Gest, Jest, Kest

Examples

        a <- allstats(swedishpines,dataname="Swedish Pines")
        if(interactive()) {
        plot(a)
        plot(a, subset=list("r<=15","r<=15","r<=15","r<=50"))
        }

[Package spatstat.explore version 3.3-4 Index]