as.data.frame.envelope {spatstat.explore}	R Documentation

Coerce Envelope to Data Frame

Description

Converts an envelope object to a data frame.

Usage

## S3 method for class 'envelope'
as.data.frame(x, ..., simfuns=FALSE)

Arguments

x
Envelope object (class "envelope").

...
Ignored.

simfuns
Logical value indicating whether the result should include the values of the simulated functions that were used to build the envelope.

Details

This is a method for the generic function as.data.frame for the class of envelopes (see envelope.

The result is a data frame with columns containing the values of the function argument (usually named r), the function estimate for the original point pattern data (obs), the upper and lower envelope limits (hi and lo), and possibly additional columns.

If simfuns=TRUE, the result also includes columns of values of the simulated functions that were used to compute the envelope. This is possible only when the envelope was computed with the argument savefuns=TRUE in the call to envelope.

Value

A data frame.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

  E <- envelope(cells, nsim=5, savefuns=TRUE)
  tail(as.data.frame(E))
  tail(as.data.frame(E, simfuns=TRUE))

[Package spatstat.explore version 3.3-4 Index]