as.tess {spatstat.explore}	R Documentation

Convert Data To Tessellation

Description

Converts data specifying a tessellation, in any of several formats, into an object of class "tess".

Usage

 ## S3 method for class 'quadrattest'
as.tess(X)

Arguments

X
Data to be converted to a tessellation.

Details

A tessellation is a collection of disjoint spatial regions (called tiles) that fit together to form a larger spatial region. This command creates an object of class "tess" that represents a tessellation.

This function converts data in any of several formats into an object of class "tess" for use by the spatstat package. The argument X may be

an object of class "tess". The object will be stripped of any extraneous attributes and returned.

a pixel image (object of class "im") with pixel values that are logical or factor values. Each level of the factor will determine a tile of the tessellation.

a window (object of class "owin"). The result will be a tessellation consisting of a single tile.

a set of quadrat counts (object of class "quadratcount") returned by the command quadratcount. The quadrats used to generate the counts will be extracted and returned as a tessellation.

a quadrat test (object of class "quadrattest") returned by the command quadrat.test. The quadrats used to perform the test will be extracted and returned as a tessellation.

a list of windows (objects of class "owin") giving the tiles of the tessellation.

The function as.tess is generic, with methods for various classes, as listed above.

Value

An object of class "tess" specifying a tessellation.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

tess

Examples

 h <- quadrat.test(nztrees, nx=4, ny=3)
 as.tess(h)

[Package spatstat.explore version 3.3-4 Index]