auc {spatstat.explore}	R Documentation

Area Under ROC Curve

Description

Compute the AUC (area under the Receiver Operating Characteristic curve) for an observed point pattern.

Usage

auc(X, ...)

## S3 method for class 'ppp'
auc(X, covariate, ..., high = TRUE)

Arguments

X
Point pattern (object of class "ppp" or "lpp") or fitted point process model (object of class "ppm", "kppm", "slrm" or "lppm").

covariate
Spatial covariate. Either a function(x,y), a pixel image (object of class "im"), or one of the strings "x" or "y" indicating the Cartesian coordinates.

high
Logical value indicating whether the threshold operation should favour high or low values of the covariate.

...
Arguments passed to as.mask controlling the pixel resolution for calculations.

Details

This command computes the AUC, the area under the Receiver Operating Characteristic curve. The ROC itself is computed by roc.

For a point pattern X and a covariate Z, the AUC is a numerical index that measures the ability of the covariate to separate the spatial domain into areas of high and low density of points. Let x[i] be a randomly-chosen data point from X and U a randomly-selected location in the study region. The AUC is the probability that Z(x[i]) > Z(U) assuming high=TRUE. That is, AUC is the probability that a randomly-selected data point has a higher value of the covariate Z than does a randomly-selected spatial location. The AUC is a number between 0 and 1. A value of 0.5 indicates a complete lack of discriminatory power.

Value

Numeric. For auc.ppp and auc.lpp, the result is a single number giving the AUC value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Lobo, J.M., Jimenez-Valverde, A. and Real, R. (2007) AUC: a misleading measure of the performance of predictive distribution models. Global Ecology and Biogeography 17(2) 145–151.

Nam, B.-H. and D'Agostino, R. (2002) Discrimination index, the area under the ROC curve. Pages 267–279 in Huber-Carol, C., Balakrishnan, N., Nikulin, M.S. and Mesbah, M., Goodness-of-fit tests and model validity, Birkhauser, Basel.

See Also

roc

Examples

  auc(swedishpines, "x")

[Package spatstat.explore version 3.3-4 Index]