bind.fv {spatstat.explore}	R Documentation

Combine Function Value Tables

Description

Advanced Use Only. Combine objects of class "fv", or glue extra columns of data onto an existing "fv" object.

Usage

## S3 method for class 'fv'
cbind(...)

bind.fv(x, y, labl = NULL, desc = NULL, preferred = NULL, clip=FALSE)

Arguments

...
Any number of arguments, which are objects of class "fv", or other data. See Details.

x
An object of class "fv".

y
Either an object of class "fv", a data frame, or a function. See Details.

labl
Plot labels (see fv) for columns of y. A character vector.

desc
Descriptions (see fv) for columns of y. A character vector.

preferred
Character string specifying the column which is to be the new recommended value of the function.

clip
Logical value indicating whether each object must have exactly the same domain, that is, the same sequence of values of the function argument (clip=FALSE, the default) or whether objects with different domains are permissible and will be restricted to a common domain (clip=TRUE).

Details

This documentation is provided for experienced programmers who want to modify the internal behaviour of spatstat.

The function cbind.fv is a method for the generic R function cbind. It combines any number of objects of class "fv" into a single object of class "fv". The objects must be compatible, in the sense that they have identical values of the function argument.

The function bind.fv is a lower level utility which glues additional columns onto an existing object x of class "fv". It has three modes of use:

If the additional dataset y is an object of class "fv", then x and y must be compatible as described above. Then the columns of y that contain function values will be appended to the object x.

Alternatively if y is a data frame, then y must have the same number of rows as x. All columns of y will be appended to x.

Alternatively if y is a function in the R language, then this function will be evaluated at the argument values stored in the object x, and these function values will be appended as a new column to x.

The arguments labl and desc provide plot labels and description strings (as described in fv) for the new columns. If y is an object of class "fv" then labl and desc are optional, and default to the relevant entries in the object y. If y is a data frame then labl and desc should be provided, but there is a default.

For additional flexibility, cbind.fv also accepts arguments which are data frames or functions.

Value

An object of class "fv".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

fv for creating objects of class "fv" from raw data.

collapse.fv for combining several "fv" objects with similar columns.

with.fv for evaluating expressions.

fvnames for extracting and assigning the column names of standard components of "fv" objects.

Undocumented functions for modifying an "fv" object include tweak.fv.entry and rebadge.fv.

Examples

   K1 <- Kest(cells, correction="border")
   K2 <- Kest(cells, correction="iso")

   # remove column 'theo' to avoid duplication
   K2 <- K2[, names(K2) != "theo"]

   cbind(K1, K2)

   bind.fv(K1, K2, preferred="iso")

   # constrain border estimate to be monotonically increasing
   bm <- cumsum(c(0, pmax(0, diff(K1$border))))
   bind.fv(K1, data.frame(bmono=bm),
               "%s[bmo](r)",
               "monotone border-corrected estimate of %s",
               "bmono")

   # add a column of values defined by a function
   cbind(K1, upper=function(r) { pi * r^2 + 0.1 })

[Package spatstat.explore version 3.3-4 Index]