bw.frac {spatstat.explore}	R Documentation

Bandwidth Selection Based on Window Geometry

Description

Select a smoothing bandwidth for smoothing a point pattern, based only on the geometry of the spatial window. The bandwidth is a specified quantile of the distance between two independent random points in the window.

Usage

   bw.frac(X, ..., f=1/4)

Arguments

X
A window (object of class "owin") or point pattern (object of class "ppp") or other data which can be converted to a window using as.owin.

...
Arguments passed to distcdf.

f
Probability value (between 0 and 1) determining the quantile of the distribution.

Details

This function selects an appropriate bandwidth sigma for the kernel estimator of point process intensity computed by density.ppp.

The bandwidth σ is computed as a quantile of the distance between two independent random points in the window. The default is the lower quartile of this distribution.

If F(r) is the cumulative distribution function of the distance between two independent random points uniformly distributed in the window, then the value returned is the quantile with probability f. That is, the bandwidth is the value r such that F(r) = f.

The cumulative distribution function F(r) is computed using distcdf. We then we compute the smallest number r such that F(r) >= f.

Value

A numerical value giving the selected bandwidth. The result also belongs to the class "bw.frac" which can be plotted to show the cumulative distribution function and the selected quantile.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

For estimating point process intensity, see density.ppp, bw.diggle, bw.ppl, bw.scott, bw.CvL.

For other smoothing purposes, see bw.stoyan, bw.smoothppp, bw.relrisk.

Examples

  h <- bw.frac(letterR)
  h
  plot(h, main="bw.frac(letterR)")

[Package spatstat.explore version 3.3-4 Index]