clarkevans.test {spatstat.explore}	R Documentation

Clark and Evans Test

Description

Performs the Clark-Evans test of aggregation for a spatial point pattern.

Usage

clarkevans.test(X, ...,
               correction,
               clipregion=NULL,
               alternative=c("two.sided", "less", "greater",
                             "clustered", "regular"),
               method=c("asymptotic", "MonteCarlo"),
               nsim=999)

Arguments

X
A spatial point pattern (object of class "ppp").

...
Ignored.

correction
Character string. The type of edge correction to be applied. See clarkevans and Details below.

clipregion
Clipping region for the guard area correction. A window (object of class "owin"). See clarkevans

alternative
String indicating the type of alternative for the hypothesis test. Partially matched.

method
Character string (partially matched) specifying how to calculate the p-value of the test. See Details.

nsim
Number of Monte Carlo simulations to perform, if a Monte Carlo p-value is required.

Details

This command uses the Clark and Evans (1954) aggregation index R as the basis for a crude test of clustering or ordering of a point pattern.

The Clark-Evans aggregation index R is computed by the separate function clarkevans.

This command clarkevans.text performs a hypothesis test of clustering or ordering of the point pattern X based on the Clark-Evans index R. The null hypothesis is Complete Spatial Randomness, i.e.\ a uniform Poisson process. The alternative hypothesis is specified by the argument alternative:

alternative="less" or alternative="clustered": the alternative hypothesis is that R < 1 corresponding to a clustered point pattern;

alternative="greater" or alternative="regular": the alternative hypothesis is that R > 1 corresponding to a regular or ordered point pattern;

alternative="two.sided": the alternative hypothesis is that R != 1 corresponding to a clustered or regular pattern.

The Clark-Evans index R is first computed for the point pattern dataset X using the edge correction determined by the arguments correction and clipregion. These arguments are documented in the help file for clarkevans.

If method="asymptotic" (the default), the p-value for the test is computed by standardising R as proposed by Clark and Evans (1954) and referring the standardised statistic to the standard Normal distribution. For this asymptotic test, the default edge correction is correction="Donnelly" if the window of X is a rectangle, and correction="cdf" otherwise. It is strongly recommended to avoid using correction="none" which would lead to a severely biased test.

If method="MonteCarlo", the p-value for the test is computed by comparing the observed value of R to the results obtained from nsim simulated realisations of Complete Spatial Randomness conditional on the observed number of points. This test is theoretically exact for any choice of edge correction, but may have lower power than the asymptotic test. For this Monte Carlo test, the default edge correction is correction="none" for computational efficiency.

Value

An object of class "htest" representing the result of the test.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Clark, P.J. and Evans, F.C. (1954) Distance to nearest neighbour as a measure of spatial relationships in populations. Ecology 35, 445–453.

Donnelly, K. (1978) Simulations to determine the variance and edge-effect of total nearest neighbour distance. In Simulation methods in archaeology, Cambridge University Press, pp 91–95.

See Also

clarkevans, hopskel.test

Examples

  # Redwood data - clustered
  clarkevans.test(redwood)
  clarkevans.test(redwood, alternative="clustered")
  clarkevans.test(redwood, correction="cdf", method="MonteCarlo", nsim=39)

[Package spatstat.explore version 3.3-4 Index]