collapse.fv {spatstat.explore}	R Documentation

Collapse Several Function Tables into One

Description

Combines several function tables (objects of class "fv") into a single function table, merging columns that are identical and relabelling columns that are different.

Usage

## S3 method for class 'fv'
collapse(object, ..., same = NULL, different = NULL)

## S3 method for class 'anylist'
collapse(object, ..., same = NULL, different = NULL)

Arguments

object
An object of class "fv", or a list of such objects.

...
Additional objects of class "fv".

same
Character string or character vector specifying a column or columns of function values that are identical in different "fv" objects. These columns will be included only once in the result.

different
Character string or character vector specifying a column or columns of function values, that are different in different "fv" objects. Each of these columns of data will be included, with labels that distinguish them from each other.

Details

This is a method for the generic function collapse.

It combines the data in several function tables (objects of class "fv", see fv.object) to make a single function table. It is essentially a smart wrapper for cbind.fv.

A typical application is to calculate the same summary statistic (such as the K function) for different point patterns, and then to use collapse.fv to combine the results into a single object that can easily be plotted. See the Examples.

The arguments object and ... should be function tables (objects of class "fv", see fv.object) that are compatible in the sense that they have the same values of the function argument. (This can be ensured by applying harmonise.fv to them.)

The argument same identifies any columns that are present in some or all of the function tables, and which are known to contain exactly the same values in each table that includes them. This column or columns will be included only once in the result.

The argument different identifies any columns that are present in some or all of the function tables, and which may contain different numerical values in different tables. Each of these columns will be included, with labels to distinguish them.

Columns that are not named in same or different will not be included.

The function argument is always included and does not need to be specified.

The arguments same and different can be NULL, or they can be character vectors containing the names of columns of object. The argument different can be one of the abbreviations recognised by fvnames.

Value

Object of class "fv".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

fv.object, cbind.fv

Examples

  # generate simulated data
  X <- replicate(3, rpoispp(100), simplify=FALSE)
  names(X) <- paste("Simulation", 1:3)
  # compute K function estimates
  Klist <- anylapply(X, Kest)
  # collapse
  K <- collapse(Klist, same="theo", different="iso")
  K

[Package spatstat.explore version 3.3-4 Index]