compatible.fasp {spatstat.explore}	R Documentation

Test Whether Function Arrays Are Compatible

Description

Tests whether two or more function arrays (class "fasp") are compatible.

Usage

  ## S3 method for class 'fasp'
compatible(A, B, ...)

Arguments

A,B,...
Two or more function arrays (object of class "fasp").

Details

An object of class "fasp" can be regarded as an array of functions. Such objects are returned by the command alltypes.

This command tests whether such objects are compatible (so that, for example, they could be added or subtracted). It is a method for the generic command compatible.

The function arrays are compatible if the arrays have the same dimensions, and the corresponding elements in each cell of the array are compatible as defined by compatible.fv.

Value

Logical value: TRUE if the objects are compatible, and FALSE if they are not.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

eval.fasp

[Package spatstat.explore version 3.3-4 Index]