compatible.fv {spatstat.explore}	R Documentation

Test Whether Function Objects Are Compatible

Description

Tests whether two or more function objects (class "fv") are compatible.

Usage

  ## S3 method for class 'fv'
compatible(A, B, ..., samenames=TRUE)

Arguments

A,B,...
Two or more function value objects (class "fv").

samenames
Logical value indicating whether to check for complete agreement between the column names of the objects (samenames=TRUE, the default) or just to check that the name of the function argument is the same (samenames=FALSE).

Details

An object of class "fv" is essentially a data frame containing several different statistical estimates of the same function. Such objects are returned by Kest and its relatives.

This command tests whether such objects are compatible (so that, for example, they could be added or subtracted). It is a method for the generic command compatible.

The functions are compatible if they have been evaluated at the same sequence of values of the argument r, and if the statistical estimates have the same names.

Value

Logical value: TRUE if the objects are compatible, and FALSE if they are not.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

eval.fv

[Package spatstat.explore version 3.3-4 Index]