densityHeat {spatstat.explore}	R Documentation

Diffusion Estimate of Point Pattern Intensity

Description

Computes a diffusion estimate of intensity for a point pattern.

Usage

densityHeat(x, sigma, ...)

Arguments

x
Point pattern (object of class "ppp" or another class).

sigma
Smoothing bandwidth. Usually a single number giving the equivalent standard deviation of the smoother.

...
Additional arguments depending on the method.

Details

The generic function densityHeat computes an estimate of point process intensity using a diffusion kernel method.

Further details depend on the class of point pattern x. See the help file for the appropriate method.

Value

Depends on the class of x.

Author(s)

Adrian Baddeley and Tilman Davies.

See Also

For two-dimensional point patterns (objects of class "ppp"), the diffusion kernel estimator is densityHeat.ppp. The usual kernel estimator is density.ppp, and the tessellation-based estimator is adaptive.density.

[Package spatstat.explore version 3.3-4 Index]