dimhat {spatstat.explore}	R Documentation

Estimate Dimension of Central Subspace

Description

Given the kernel matrix that characterises a central subspace, this function estimates the dimension of the subspace.

Usage

  dimhat(M)

Arguments

M
Kernel of subspace. A symmetric, non-negative definite, numeric matrix, typically obtained from sdr.

Details

This function computes the maximum descent estimate of the dimension of the central subspace with a given kernel matrix M.

The matrix M should be the kernel matrix of a central subspace, which can be obtained from sdr. It must be a symmetric, non-negative-definite, numeric matrix.

The algorithm finds the eigenvalues lambda[1] ≥ ...≥ lambda[n] of M, and then determines the index k for which lambda[k]/lambda[k-1] is greatest.

Value

A single integer giving the estimated dimension.

Author(s)

Matlab original by Yongtao Guan, translated to R by Suman Rakshit.

References

Guan, Y. and Wang, H. (2010) Sufficient dimension reduction for spatial point processes directed by Gaussian random fields. Journal of the Royal Statistical Society, Series B, 72, 367–387.

See Also

sdr, subspaceDistance

[Package spatstat.explore version 3.3-4 Index]