envelope.envelope {spatstat.explore}	R Documentation

Recompute Envelopes

Description

Given a simulation envelope (object of class "envelope"), compute another envelope from the same simulation data using different parameters.

Usage

## S3 method for class 'envelope'
envelope(Y, fun = NULL, ...,
                            transform=NULL, global=FALSE, VARIANCE=FALSE)

Arguments

Y
A simulation envelope (object of class "envelope").

fun
Optional. Summary function to be applied to the simulated point patterns.

...,transform,global,VARIANCE
Parameters controlling the type of envelope that is re-computed. See envelope.

Details

This function can be used to re-compute a simulation envelope from previously simulated data, using different parameter settings for the envelope: for example, a different significance level, or a global envelope instead of a pointwise envelope.

The function envelope is generic. This is the method for the class "envelope".

The argument Y should be a simulation envelope (object of class "envelope") produced by any of the methods for envelope. Additionally, Y must contain either

the simulated point patterns that were used to create the original envelope (so Y should have been created by calling envelope with savepatterns=TRUE);

the summary functions of the simulated point patterns that were used to create the original envelope (so Y should have been created by calling envelope with savefuns=TRUE).

If the argument fun is given, it should be a summary function that can be applied to the simulated point patterns that were used to create Y. The envelope of the summary function fun for these point patterns will be computed using the parameters specified in ....

If fun is not given, then:

If Y contains the summary functions that were used to compute the original envelope, then the new envelope will be computed from these original summary functions.

Otherwise, if Y contains the simulated point patterns. then the K function Kest will be applied to each of these simulated point patterns, and the new envelope will be based on the K functions.

The new envelope will be computed using the parameters specified in ....

See envelope for a full list of envelope parameters. Frequently-used parameters include nrank and nsim (to change the number of simulations used and the significance level of the envelope), global (to change from pointwise to global envelopes) and VARIANCE (to compute the envelopes from the sample moments instead of the ranks).

Value

An envelope (object of class "envelope".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

envelope

Examples

  E <- envelope(cells, Kest, nsim=19, savefuns=TRUE, savepatterns=TRUE)
  E2 <- envelope(E, nrank=2)
  Eg <- envelope(E, global=TRUE)
  EG <- envelope(E, Gest)
  EL <- envelope(E, transform=expression(sqrt(./pi)))

[Package spatstat.explore version 3.3-4 Index]