fasp.object {spatstat.explore}	R Documentation

Function Arrays for Spatial Patterns

Description

A class "fasp" to represent a “matrix” of functions, amenable to plotting as a matrix of plot panels.

Details

An object of this class is a convenient way of storing (and later plotting, editing, etc) a set of functions f[i,j](r) of a real argument r, defined for each possible pair (i,j) of indices 1 <= i,j <= n. We may think of this as a matrix or array of functions f[i,j].

Function arrays are particularly useful in the analysis of a multitype point pattern (a point pattern in which the points are identified as belonging to separate types). We may want to compute a summary function for the points of type i only, for each of the possible types i. This produces a 1 * m array of functions. Alternatively we may compute a summary function for each possible pair of types (i,j). This produces an m * m array of functions.

For multitype point patterns the command alltypes will compute arrays of summary functions for each possible type or for each possible pair of types. The function alltypes returns an object of class "fasp".

An object of class "fasp" is a list containing at least the following components:

fns

A list of data frames, each representing one of the functions.

which

A matrix representing the spatial arrangement of the functions. If which[i,j] = k then the function represented by fns[[k]] should be plotted in the panel at position (i,j). If which[i,j] = NA then nothing is plotted in that position.

titles

A list of character strings, providing suitable plotting titles for the functions.

default.formulae

A list of default formulae for plotting each of the functions.

title

A character string, giving a default title for the array when it is plotted.

Functions available

There are methods for plot, print and "[" for this class.

The plot method displays the entire array of functions. The method [.fasp selects a sub-array using the natural indices i,j.

The command eval.fasp can be used to apply a transformation to each function in the array, and to combine two arrays.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

alltypes, plot.fasp, [.fasp, eval.fasp

Examples

  GG <- alltypes(amacrine, 'G')

  plot(GG)

  # select the row corresponding to cells of type "on"
  Gon <- GG["on", ]
  plot(Gon)

  # extract the G function for i = "on", j = "off"
  Gonoff <- GG["on", "off", drop=TRUE]

  # Fisher variance stabilising transformation
  GGfish <- eval.fasp(asin(sqrt(GG)))
  plot(GGfish)

[Package spatstat.explore version 3.3-4 Index]