formula.fv {spatstat.explore}	R Documentation

Extract or Change the Plot Formula for a Function Value Table

Description

Extract or change the default plotting formula for an object of class "fv" (function value table).

Usage

## S3 method for class 'fv'
formula(x, ...)

formula(x, ...) <- value

## S3 replacement method for class 'fv'
formula(x, ...) <- value

Arguments

x
An object of class "fv", containing the values of several estimates of a function.

...
Arguments passed to other methods.

value
New value of the formula. Either a formula or a character string.

Details

A function value table (object of class "fv", see fv.object) is a convenient way of storing and plotting several different estimates of the same function.

The default behaviour of plot(x) for a function value table x is determined by a formula associated with x called its plot formula. See plot.fv for explanation about these formulae.

The function formula.fv is a method for the generic command formula. It extracts the plot formula associated with the object.

The function formula<- is generic. It changes the formula associated with an object.

The function formula<-.fv is the method for formula<- for the class "fv". It changes the plot formula associated with the object.

Value

The result of formula.fv is a character string containing the plot formula. The result of formula<-.fv is a new object of class "fv".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

fv, plot.fv, formula.

Examples

   K <- Kest(cells)
   formula(K)
   formula(K) <- (iso ~ r)

[Package spatstat.explore version 3.3-4 Index]