fv.object {spatstat.explore}	R Documentation

Function Value Table

Description

A class "fv" to support the convenient plotting of several estimates of the same function.

Details

An object of this class is a convenient way of storing and plotting several different estimates of the same function.

It is a data frame with extra attributes indicating the recommended way of plotting the function, and other information.

There are methods for print and plot for this class.

Objects of class "fv" are returned by Fest, Gest,Jest, and Kest along with many other functions.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

Objects of class "fv" are returned by Fest, Gest,Jest, and Kest along with many other functions.

See plot.fv for plotting an "fv" object.

See as.function.fv to convert an "fv" object to an R function.

Use cbind.fv to combine several "fv" objects. Use bind.fv to glue additional columns onto an existing "fv" object.

Undocumented functions for modifying an "fv" object include fvnames, fvnames<-, tweak.fv.entry and rebadge.fv.

Examples

    K <- Kest(cells)
    class(K)

    K  # prints a sensible summary

    plot(K)

[Package spatstat.explore version 3.3-4 Index]