fvnames {spatstat.explore}	R Documentation

Abbreviations for Groups of Columns in Function Value Table

Description

Groups of columns in a function value table (object of class "fv") identified by standard abbreviations.

Usage

fvnames(X, a = ".")

fvnames(X, a = ".") <- value

Arguments

X
Function value table (object of class "fv"). See fv.object.

a
One of the standard abbreviations listed below.

value
Character vector containing names of columns of X.

Details

An object of class "fv" represents a table of values of a function, usually a summary function for spatial data such as the K-function, for which several different statistical estimators may be available. The different estimates are stored as columns of the table.

Auxiliary information carried in the object X specifies some columns or groups of columns of this table that should be used for particular purposes. For convenience these groups can be referred to by standard abbreviations which are recognised by various functions in the spatstat package, such as plot.fv.

These abbreviations are:

".x" 	 the function argument

".y" 	 the recommended value of the function

"." 	 all function values to be plotted by default

(in order of plotting)

".s" 	 the upper and lower limits of shading

(for envelopes and confidence intervals)

".a" 	 all function values (in column order)

The command fvnames(X, a) expands the abbreviation a and returns a character vector containing the names of the columns.

The assignment fvnames(X, a) <- value changes the definition of the abbreviation a to the character string value (which should be the name of another column of X). The column names of X are not changed.

Note that fvnames(x, ".") lists the columns of values that will be plotted by default, in the order that they would be plotted, not in order of the column position. The order in which curves are plotted affects the colours and line styles associated with the curves.

Value

For fvnames, a character vector.

For fvnames<-, the updated object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

fv.object, plot.fv

Examples

   K <- Kest(cells)
   fvnames(K, ".y")
   fvnames(K, ".y") <- "trans"

[Package spatstat.explore version 3.3-4 Index]