increment.fv {spatstat.explore}	R Documentation

Increments of a Function

Description

Compute the change in the value of a function f when the function argument increases by delta.

Usage

increment.fv(f, delta)

Arguments

f
Object of class "fv" representing a function.

delta
Numeric. The increase in the value of the function argument.

Details

This command computes the new function

g(x) = f(x+h) - f(x-h)

where h = delta/2. The value of g(x) is the change in the value of f over an interval of length delta centred at x.

Value

Another object of class "fv" compatible with X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

fv.object, deriv.fv

Examples

  plot(increment.fv(Kest(cells), 0.05))

[Package spatstat.explore version 3.3-4 Index]