localKdot {spatstat.explore}	R Documentation

Local Multitype K Function (Dot-Type)

Description

for a multitype point pattern, computes the dot-type version of the local K function.

Usage

  localKdot(X, from, ..., rmax = NULL,
              correction = "Ripley", verbose = TRUE, rvalue=NULL)
  localLdot(X, from, ..., rmax = NULL, correction = "Ripley")

Arguments

X
A multitype point pattern (object of class "ppp" with marks which are a factor).

...
Further arguments passed from localLdot to localKdot.

rmax
Optional. Maximum desired value of the argument r.

from
Type of points from which distances should be measured. A single value; one of the possible levels of marks(X), or an integer indicating which level.

correction
String specifying the edge correction to be applied. Options are "none", "translate", "translation", "Ripley", "isotropic" or "best". Only one correction may be specified.

verbose
Logical flag indicating whether to print progress reports during the calculation.

rvalue
Optional. A single value of the distance argument r at which the function L or K should be computed.

Details

Given a multitype spatial point pattern X, the local dot-type K function localKdot is the local version of the multitype K function Kdot. Recall that Kdot(X, from) is a sum of contributions from all pairs of points in X where the first point belongs to from. The local dot-type K function is defined for each point X[i] that belongs to type from, and it consists of all the contributions to the dot-type K function that originate from point X[i]:

K[i,from,to](r) = sqrt( (a/((n-1)* pi)) * sum[j] e[i,j])

where the sum is over all points j != i that lie within a distance r of the ith point, a is the area of the observation window, n is the number of points in X, and e[i,j] is an edge correction term (as described in Kest). The value of K[i,from](r) can also be interpreted as one of the summands that contributes to the global estimate of the Kdot function.

By default, the function K[i,from](r) is computed for a range of r values for each point i belonging to type from. The results are stored as a function value table (object of class "fv") with a column of the table containing the function estimates for each point of the pattern X belonging to type from.

Alternatively, if the argument rvalue is given, and it is a single number, then the function will only be computed for this value of r, and the results will be returned as a numeric vector, with one entry of the vector for each point of the pattern X belonging to type from.

The local dot-type L function localLdot is computed by applying the transformation L(r) = sqrt(K(r)/(2*pi)).

Value

If rvalue is given, the result is a numeric vector of length equal to the number of points in the point pattern that belong to type from.

If rvalue is absent, the result is an object of class "fv", see fv.object, which can be plotted directly using plot.fv. Essentially a data frame containing columns

r
the vector of values of the argument r at which the function K has been estimated

theo
the theoretical value K(r) = pi * r^2 or L(r)=r for a stationary Poisson process

together with columns containing the values of the neighbourhood density function for each point in the pattern. Column i corresponds to the ith point of type from. The last two columns contain the r and theo values.

Author(s)

Ege Rubak rubak@math.aau.dk and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

Kdot, Ldot, localK, localL.

Examples

  X <- amacrine

  # compute all the local Ldot functions
  L <- localLdot(X)

  # plot all the local Ldot functions against r
  plot(L, main="local Ldot functions for amacrine", legend=FALSE)

  # plot only the local L function for point number 7
  plot(L, iso007 ~ r)
  
  # compute the values of L(r) for r = 0.1 metres
  L12 <- localLdot(X, rvalue=0.1)

[Package spatstat.explore version 3.3-4 Index]