markvario {spatstat.explore}	R Documentation

Mark Variogram

Description

Estimate the mark variogram of a marked point pattern.

Usage

markvario(X, correction = c("isotropic", "Ripley", "translate"),
r = NULL, method = "density", ..., normalise=FALSE)

Arguments

X
The observed point pattern. An object of class "ppp" or something acceptable to as.ppp. It must have marks which are numeric.

correction
A character vector containing any selection of the options "isotropic", "Ripley" or "translate". It specifies the edge correction(s) to be applied.

r
numeric vector. The values of the argument r at which the mark variogram gamma(r) should be evaluated. There is a sensible default.

method
A character vector indicating the user's choice of density estimation technique to be used. Options are "density", "loess", "sm" and "smrep".

...
Other arguments passed to markcorr, or passed to the density estimation routine (density, loess or sm.density) selected by method.

normalise
If TRUE, normalise the variogram by dividing it by the estimated mark variance.

Details

The mark variogram gamma(r) of a marked point process X is a measure of the dependence between the marks of two points of the process a distance r apart. It is informally defined as

gamma(r) = E[(1/2) * (M1 - M2)^2 ]

where E[ ] denotes expectation and M1,M2 are the marks attached to two points of the process a distance r apart.

The mark variogram of a marked point process is analogous, but not equivalent, to the variogram of a random field in geostatistics. See Waelder and Stoyan (1996).

Value

An object of class "fv" (see fv.object).

Essentially a data frame containing numeric columns

r
the values of the argument r at which the mark variogram gamma(r) has been estimated

theo
the theoretical value of gamma(r) when the marks attached to different points are independent; equal to the sample variance of the marks

together with a column or columns named "iso" and/or "trans", according to the selected edge corrections. These columns contain estimates of the function gamma(r) obtained by the edge corrections named.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Cressie, N.A.C. (1991) Statistics for spatial data. John Wiley and Sons, 1991.

Mase, S. (1996) The threshold method for estimating annual rainfall. Annals of the Institute of Statistical Mathematics 48 (1996) 201-213.

Waelder, O. and Stoyan, D. (1996) On variograms in point process statistics. Biometrical Journal 38 (1996) 895-905.

See Also

Mark correlation function markcorr for numeric marks.

Mark connection function markconnect and multitype K-functions Kcross, Kdot for factor-valued marks.

Examples

    # Longleaf Pine data
    # marks represent tree diameter
    # Subset of this large pattern
    swcorner <- owin(c(0,100),c(0,100))
    sub <- longleaf[ , swcorner]
    # mark correlation function
    mv <- markvario(sub)
    plot(mv)

[Package spatstat.explore version 3.3-4 Index]