methods.rho2hat {spatstat.explore}	R Documentation

Methods for Intensity Functions of Two Spatial Covariates

Description

These are methods for the class "rho2hat".

Usage

## S3 method for class 'rho2hat'
plot(x, ..., do.points=FALSE)

## S3 method for class 'rho2hat'
print(x, ...)

## S3 method for class 'rho2hat'
predict(object, ..., relative=FALSE)

Arguments

x,object
An object of class "rho2hat".

...
Arguments passed to other methods.

do.points
Logical value indicating whether to plot the observed values of the covariates at the data points.

relative
Logical value indicating whether to compute the estimated point process intensity (relative=FALSE) or the relative risk (relative=TRUE) in the case of a relative risk estimate.

Details

These functions are methods for the generic commands print, predict and plot for the class "rho2hat".

An object of class "rho2hat" is an estimate of the intensity of a point process, as a function of two given spatial covariates. See rho2hat.

The method plot.rho2hat displays the estimated function rho using plot.fv, and optionally adds a rug plot of the observed values of the covariate. In this plot the two axes represent possible values of the two covariates.

The method predict.rho2hat computes a pixel image of the intensity rho(Z1(u), Z2(u)) at each spatial location u, where Z1(u) and Z2(u) are the two spatial covariates.

Value

For predict.rho2hat the value is a pixel image (object of class "im"). For other functions, the value is NULL.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

rho2hat

Examples

  r2 <- with(bei.extra, rho2hat(bei, elev, grad))
  r2
  plot(r2)
  plot(predict(r2))

[Package spatstat.explore version 3.3-4 Index]