methods.ssf {spatstat.explore}	R Documentation

Methods for Spatially Sampled Functions

Description

Methods for various generic commands, for the class "ssf" of spatially sampled functions.

Usage

  ## S3 method for class 'ssf'
marks(x, ...)

  ## S3 replacement method for class 'ssf'
marks(x, ...) <- value

  ## S3 method for class 'ssf'
unmark(X)

  ## S3 method for class 'ssf'
as.im(X, ...)

  ## S3 method for class 'ssf'
as.function(x, ...)

  ## S3 method for class 'ssf'
as.ppp(X, ...)

  ## S3 method for class 'ssf'
print(x, ..., brief=FALSE)

  ## S3 method for class 'ssf'
summary(object, ...)

  ## S3 method for class 'ssf'
range(x, ...)

  ## S3 method for class 'ssf'
min(x, ...)

  ## S3 method for class 'ssf'
max(x, ...)

  ## S3 method for class 'ssf'
integral(f, domain=NULL, ..., weights=attr(f, "weights"))

Arguments

x,X,f,object
A spatially sampled function (object of class "ssf").

...
Arguments passed to the default method.

brief
Logical value controlling the amount of detail printed.

value
Matrix of replacement values for the function.

domain
Optional. Domain of integration. An object of class"owin" or "tess".

weights
Optional. Numeric vector of quadrature weights associated with the sample points.

Details

An object of class "ssf" represents a function (real- or vector-valued) that has been sampled at a finite set of points.

The commands documented here are methods for this class, for the generic commands marks, marks<-, unmark, as.im, as.function, as.ppp, print, summary, range, min, max and integral.

Value

marks returns a matrix.

marks(x) <- value returns an object of class "ssf".

as.owin returns a window (object of class "owin").

as.ppp and unmark return a point pattern (object of class "ppp").

as.function returns a function(x,y) of class "funxy".

print returns NULL.

summary returns an object of class "summary.ssf" which has a print method.

range returns a numeric vector of length 2. min and max return a single numeric value.

integral returns a numeric or complex value, vector, or matrix. integral(f) returns a numeric or complex value (if f had numeric or complex values) or a numeric vector (if f had vector values). If domain is a tessellation then integral(f, domain) returns a numeric or complex vector with one entry for each tile (if f had numeric or complex values) or a numeric matrix with one row for each tile (if f had vector values).

Author(s)

Adrian Baddeley

See Also

ssf

Examples

  g <- distfun(cells[1:4])
  X <- rsyst(Window(cells), 10)
  f <- ssf(X, g(X))
  f
  summary(f)
  marks(f)
  as.ppp(f)
  as.im(f)
  integral(f)
  integral(f, quadrats(Window(f), 3))

[Package spatstat.explore version 3.3-4 Index]