miplot {spatstat.explore}	R Documentation

Morisita Index Plot

Description

Displays the Morisita Index Plot of a spatial point pattern.

Usage

miplot(X, ...)

Arguments

X
A point pattern (object of class "ppp") or something acceptable to as.ppp.

...
Optional arguments to control the appearance of the plot.

Details

Morisita (1959) defined an index of spatial aggregation for a spatial point pattern based on quadrat counts. The spatial domain of the point pattern is first divided into Q subsets (quadrats) of equal size and shape. The numbers of points falling in each quadrat are counted. Then the Morisita Index is computed as

MI = Q * sum(n[i] (n[i]-1))/(N(N-1))

where n[i] is the number of points falling in the i-th quadrat, and N is the total number of points. If the pattern is completely random, MI should be approximately equal to 1. Values of MI greater than 1 suggest clustering.

The Morisita Index plot is a plot of the Morisita Index MI against the linear dimension of the quadrats. The point pattern dataset is divided into 2 * 2 quadrats, then 3 * 3 quadrats, etc, and the Morisita Index is computed each time. This plot is an attempt to discern different scales of dependence in the point pattern data.

Value

None.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

M. Morisita (1959) Measuring of the dispersion of individuals and analysis of the distributional patterns. Memoir of the Faculty of Science, Kyushu University, Series E: Biology. 2: 215–235.

See Also

quadratcount

Examples

 miplot(longleaf)
 opa <- par(mfrow=c(2,3))
 plot(cells)
 plot(japanesepines)
 plot(redwood)
 miplot(cells)
 miplot(japanesepines)
 miplot(redwood)
 par(opa)

[Package spatstat.explore version 3.3-4 Index]