nndensity.ppp {spatstat.explore}	R Documentation

Estimate Intensity of Point Pattern Using Nearest Neighbour Distances

Description

Estimates the intensity of a point pattern using the distance from each spatial location to the kth nearest data point.

Usage

nndensity(x, ...)

## S3 method for class 'ppp'
nndensity(x, k, ..., verbose = TRUE)

Arguments

x
A point pattern (object of class "ppp") or some other spatial object.

k
Integer. The distance to the kth nearest data point will be computed. There is a sensible default.

...
Arguments passed to nnmap and as.mask controlling the pixel resolution.

verbose
Logical. If TRUE, print the value of k when it is automatically selected. If FALSE, remain silent.

Details

This function computes a quick estimate of the intensity of the point process that generated the point pattern x.

For each spatial location s, let d(s) be the distance from s to the k-th nearest point in the dataset x. If the data came from a homogeneous Poisson process with intensity lambda, then pi * d(s)^2 would follow a negative exponential distribution with mean 1/lambda, and the maximum likelihood estimate of lambda would be 1/(pi * d(s)^2). This is the estimate computed by nndensity, apart from an edge effect correction.

See Cressie (1991, equation (8.5.14), p. 654) and Silverman (1986, p. 96).

This estimator of intensity is relatively fast to compute, and is spatially adaptive (so that it can handle wide variation in the intensity function). However, it implicitly assumes the points are independent, so it does not perform well if the pattern is strongly clustered or strongly inhibited.

In normal use, the value of k should be at least 3. (Theoretically the estimator has infinite expected value if k=1, and infinite variance if k=2. The computed intensity estimate will have infinite peaks around each data point if k = 1.) The default value of k is the square root of the number of points in x, which seems to work well in many cases.

The window of x is digitised using as.mask and the values d(s) are computed using nnmap. To control the pixel resolution, see as.mask.

Value

A pixel image (object of class "im") giving the estimated intensity of the point process at each spatial location. Pixel values are intensities (number of points per unit area).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

References

Cressie, N.A.C. (1991) Statistics for spatial data. John Wiley and Sons, New York.

Silverman, B.W. (1986) Density Estimation. Chapman and Hall, New York.

See Also

density.ppp, intensity.ppp for alternative estimates of point process intensity.

Examples

   plot(nndensity(swedishpines))

[Package spatstat.explore version 3.3-4 Index]