pcf3est {spatstat.explore}	R Documentation

Pair Correlation Function of a Three-Dimensional Point Pattern

Description

Estimates the pair correlation function from a three-dimensional point pattern.

Usage

   pcf3est(X, ..., rmax = NULL, nrval = 128,
           correction = c("translation", "isotropic"),
           delta=NULL, adjust=1, biascorrect=TRUE)

Arguments

X
Three-dimensional point pattern (object of class "pp3").

...
Ignored.

rmax
Optional. Maximum value of argument r for which g3(r) will be estimated.

nrval
Optional. Number of values of r for which g3(r) will be estimated.

correction
Optional. Character vector specifying the edge correction(s) to be applied. See Details.

delta
Optional. Half-width of the Epanechnikov smoothing kernel.

adjust
Optional. Adjustment factor for the default value of delta.

biascorrect
Logical value. Whether to correct for underestimation due to truncation of the kernel near r=0.

Details

For a stationary point process Phi in three-dimensional space, the pair correlation function is

g3(r) = K3'(r)/(4 * pi * r^2)

where K3' is the derivative of the three-dimensional K-function (see K3est).

The three-dimensional point pattern X is assumed to be a partial realisation of a stationary point process Phi. The distance between each pair of distinct points is computed. Kernel smoothing is applied to these distance values (weighted by an edge correction factor) and the result is renormalised to give the estimate of g3(r).

The available edge corrections are:

"translation":

the Ohser translation correction estimator (Ohser, 1983; Baddeley et al, 1993)

"isotropic":

the three-dimensional counterpart of Ripley's isotropic edge correction (Ripley, 1977; Baddeley et al, 1993).

Kernel smoothing is performed using the Epanechnikov kernel with half-width delta. If delta is missing, the default is to use the rule-of-thumb delta = 0.26/lambda^(1/3) where lambda = n/v is the estimated intensity, computed from the number n of data points and the volume v of the enclosing box. This default value of delta is multiplied by the factor adjust.

The smoothing estimate of the pair correlation g3(r) is typically an underestimate when r is small, due to truncation of the kernel at r=0. If biascorrect=TRUE, the smoothed estimate is approximately adjusted for this bias. This is advisable whenever the dataset contains a sufficiently large number of points.

Value

A function value table (object of class "fv") that can be plotted, printed or coerced to a data frame containing the function values.

Additionally the value of delta is returned as an attribute of this object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rana Moyeed.

References

Baddeley, A.J, Moyeed, R.A., Howard, C.V. and Boyde, A. (1993) Analysis of a three-dimensional point pattern with replication. Applied Statistics 42, 641–668.

Ohser, J. (1983) On estimators for the reduced second moment measure of point processes. Mathematische Operationsforschung und Statistik, series Statistics, 14, 63 – 71.

Ripley, B.D. (1977) Modelling spatial patterns (with discussion). Journal of the Royal Statistical Society, Series B, 39, 172 – 212.

See Also

pp3 to create a three-dimensional point pattern (object of class "pp3").

F3est, G3est, K3est for other summary functions of a three-dimensional point pattern.

pcf to estimate the pair correlation function of point patterns in two dimensions or other spaces.

Examples

  X <- rpoispp3(250)
  Z <- pcf3est(X)
  Zbias <- pcf3est(X, biascorrect=FALSE)
  if(interactive()) {
    opa <- par(mfrow=c(1,2))
    plot(Z,     ylim.covers=c(0, 1.2))
    plot(Zbias, ylim.covers=c(0, 1.2))
    par(opa)
  }
  attr(Z, "delta")

[Package spatstat.explore version 3.3-4 Index]