plot.cdftest {spatstat.explore}	R Documentation

Plot a Spatial Distribution Test

Description

Plot the result of a spatial distribution test computed by cdf.test.

Usage

## S3 method for class 'cdftest'
plot(x, ...,
                   style=c("cdf", "PP", "QQ"),
                   lwd=par("lwd"), col=par("col"), lty=par("lty"),
                   lwd0=lwd, col0=2, lty0=2,
                   do.legend)

Arguments

x
Object to be plotted. An object of class "cdftest" produced by a method for cdf.test.

...
extra arguments that will be passed to the plotting function plot.default.

style
Style of plot. See Details.

col,lwd,lty
The width, colour and type of lines used to plot the empirical curve (the empirical distribution, or PP plot or QQ plot).

col0,lwd0,lty0
The width, colour and type of lines used to plot the reference curve (the predicted distribution, or the diagonal).

do.legend
Logical value indicating whether to add an explanatory legend. Applies only when style="cdf".

Details

This is the plot method for the class "cdftest". An object of this class represents the outcome of a spatial distribution test, computed by cdf.test, and based on either the Kolmogorov-Smirnov, Cramer-von Mises or Anderson-Darling test.

If style="cdf" (the default), the plot displays the two cumulative distribution functions that are compared by the test: namely the empirical cumulative distribution function of the covariate at the data points, and the predicted cumulative distribution function of the covariate under the model, both plotted against the value of the covariate. The Kolmogorov-Smirnov test statistic (for example) is the maximum vertical separation between the two curves.

If style="PP" then the P-P plot is drawn. The x coordinates of the plot are cumulative probabilities for the covariate under the model. The y coordinates are cumulative probabilities for the covariate at the data points. The diagonal line y=x is also drawn for reference. The Kolmogorov-Smirnov test statistic is the maximum vertical separation between the P-P plot and the diagonal reference line.

If style="QQ" then the Q-Q plot is drawn. The x coordinates of the plot are quantiles of the covariate under the model. The y coordinates are quantiles of the covariate at the data points. The diagonal line y=x is also drawn for reference. The Kolmogorov-Smirnov test statistic cannot be read off the Q-Q plot.

Value

NULL.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

cdf.test

Examples

   op <- options(useFancyQuotes=FALSE)

   plot(cdf.test(cells, "x"))

   if(require("spatstat.model")) {   
     # synthetic data: nonuniform Poisson process
     X <- rpoispp(function(x,y) { 100 * exp(x) }, win=square(1))

     # fit uniform Poisson process
     fit0 <- ppm(X ~1)

     # test covariate = x coordinate
     xcoord <- function(x,y) { x }

     # test wrong model
     k <- cdf.test(fit0, xcoord)

     # plot result of test
     plot(k, lwd0=3)

     plot(k, style="PP")

     plot(k, style="QQ")
   }

   options(op)

[Package spatstat.explore version 3.3-4 Index]