plot.envelope {spatstat.explore}	R Documentation

Plot a Simulation Envelope

Description

Plot method for the class "envelope".

Usage

 ## S3 method for class 'envelope'
plot(x, ..., main)

Arguments

x
An object of class "envelope", containing the variables to be plotted or variables from which the plotting coordinates can be computed.

main
Main title for plot.

...
Extra arguments passed to plot.fv.

Details

This is the plot method for the class "envelope" of simulation envelopes. Objects of this class are created by the command envelope.

This plot method is currently identical to plot.fv.

Its default behaviour is to shade the region between the upper and lower envelopes in a light grey colour. To suppress the shading and plot the upper and lower envelopes as curves, set shade=NULL. To change the colour of the shading, use the argument shadecol which is passed to plot.fv.

See plot.fv for further information on how to control the plot.

Value

Either NULL, or a data frame giving the meaning of the different line types and colours.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

envelope, plot.fv

Examples

   E <- envelope(cells, Kest, nsim=19)
   plot(E)
   plot(E, sqrt(./pi) ~ r)

[Package spatstat.explore version 3.3-4 Index]