plot.laslett {spatstat.explore}	R Documentation

Plot Laslett Transform

Description

Plot the result of Laslett's Transform.

Usage

## S3 method for class 'laslett'
plot(x, ...,
          Xpars = list(box = TRUE, col = "grey"),
          pointpars = list(pch = 3, cols = "blue"),
          rectpars = list(lty = 3, border = "green"))

Arguments

x
Object of class "laslett" produced by laslett representing the result of Laslett's transform.

...
Additional plot arguments passed to plot.solist.

Xpars
A list of plot arguments passed to plot.owin or plot.im to display the original region X before transformation.

pointpars
A list of plot arguments passed to plot.ppp to display the tangent points.

rectpars
A list of plot arguments passed to plot.owin to display the maximal rectangle.

Details

This is the plot method for the class "laslett".

The function laslett applies Laslett's Transform to a spatial region X and returns an object of class "laslett" representing the result of the transformation. The result is plotted by this method.

The plot function plot.solist is used to align the before-and-after pictures. See plot.solist for further options to control the plot.

Value

None.

Author(s)

Kassel Hingee and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

laslett

Examples

  b <- laslett(heather$coarse, plotit=FALSE)
  plot(b, main="Heather Data")

[Package spatstat.explore version 3.3-4 Index]