plot.quadrattest {spatstat.explore}	R Documentation

Display the result of a quadrat counting test.

Description

Given the result of a quadrat counting test, graphically display the quadrats that were used, the observed and expected counts, and the residual in each quadrat.

Usage

 ## S3 method for class 'quadrattest'
plot(x, ..., textargs=list())

Arguments

x
Object of class "quadrattest" containing the result of quadrat.test.

...
Additional arguments passed to plot.tess to control the display of the quadrats.

textargs
List of additional arguments passed to text.default to control the appearance of the text.

Details

This is the plot method for objects of class "quadrattest". Such an object is produced by quadrat.test and represents the result of a chi^2 test for a spatial point pattern.

The quadrats are first plotted using plot.tess. Then in each quadrat, the observed and expected counts and the Pearson residual are displayed as text using text.default. Observed count is displayed at top left; expected count at top right; and Pearson residual at bottom.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

quadrat.test, plot.tess, text.default, plot.quadratcount

Examples

   plot(quadrat.test(swedishpines, 3))

[Package spatstat.explore version 3.3-4 Index]