plot.studpermutest {spatstat.explore}	R Documentation

Plot a Studentised Permutation Test

Description

Plot the result of the studentised permutation test.

Usage

  ## S3 method for class 'studpermutest'
plot(x, fmla, ...,
          lty = NULL, col = NULL, lwd = NULL,
          lty.theo = NULL, col.theo = NULL, lwd.theo = NULL,
          lwd.mean = if (meanonly) 1 else NULL,
          lty.mean = lty, col.mean = col,
          separately = FALSE, meanonly = FALSE,
          main = if (meanonly) "group means" else NULL,
          xlim = NULL, ylim = NULL, ylab = NULL,
          legend = !add, legendpos = "topleft", lbox = FALSE, add = FALSE)

Arguments

x
An object of class "studpermutest" generated by studpermu.test and representing the result of a studentised permutation test for spatial point pattern data.

fmla
Plot formula used in plot.fv.

...
Additional graphical arguments passed to plot.fv.

lty,col,lwd
Line type, colour, and line width of the curves plotting the summary function for each point pattern in the original data. Either a single value or a vector of length equal to the number of point patterns.

lty.theo,col.theo,lwd.theo
Line type, colour, and line width of the curve representing the theoretical value of the summary function.

lty.mean,col.mean,lwd.mean
Line type, colour, and line width (as a multiple of lwd) of the curve representing the group mean of the summary function.

separately
Logical value indicating whether to plot each group of data in a separate panel.

meanonly
Logical value indicating whether to plot only the group means of the summary function.

main
Character string giving a main title for the plot.

xlim,ylim
Numeric vectors of length 2 giving the limits for the x and y coordinates of the plot or plots.

ylab
Character string or expression to be used for the label on the y axis.

legend
Logical value indicating whether to plot a legend explaining the meaning of each curve.

legendpos
Position of legend. See plot.fv.

lbox
Logical value indicating whether to plot a box around the plot.

add
Logical value indicating whether the plot should be added to the existing plot (add=TRUE) or whether a new frame should be created (add=FALSE, the default).

Details

This is the plot method for objects of class "studpermutest" which represent the result of a studentised permutation test applied to several point patterns. The test is performed by studpermu.test.

The plot shows the summary functions for each point pattern, coloured according to group. Optionally it can show the different groups in separate plot panels, or show only the group means in a single panel.

Value

Null.

Author(s)

Ute Hahn.

Modified for spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

studpermu.test

Examples

  np <- if(interactive()) 99 else 19
  testpyramidal <- studpermu.test(pyramidal, Neurons ~ group, nperm=np)
  plot(testpyramidal)
  plot(testpyramidal, meanonly=TRUE)
  plot(testpyramidal, col.theo=8, lwd.theo=4, lty.theo=1)
  plot(testpyramidal, . ~ pi * r^2)
  op <- par(mfrow=c(1,3))
  plot(testpyramidal, separately=TRUE)
  plot(testpyramidal, separately=TRUE, col=2, lty=1, lwd.mean=2, col.mean=4)
  par(op)

[Package spatstat.explore version 3.3-4 Index]