pool {spatstat.explore}	R Documentation

Pool Data

Description

Pool the data from several objects of the same class.

Usage

pool(...)

Arguments

...
Objects of the same type.

Details

The function pool is generic. There are methods for several classes, listed below.

pool is used to combine the data from several objects of the same type, and to compute statistics based on the combined dataset. It may be used to pool the estimates obtained from replicated datasets. It may also be used in high-performance computing applications, when the objects ... have been computed on different processors or in different batch runs, and we wish to combine them.

Value

An object of the same class as the arguments ....

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

pool.envelope, pool.fasp, pool.rat, pool.fv

[Package spatstat.explore version 3.3-4 Index]