pool.anylist {spatstat.explore}	R Documentation

Pool Data from a List of Objects

Description

Pool the data from the objects in a list.

Usage

## S3 method for class 'anylist'
pool(x, ...)

Arguments

x
A list, belonging to the class "anylist", containing objects that can be pooled.

...
Optional additional objects which can be pooled with the elements of x.

Details

The function pool is generic. Its purpose is to combine data from several objects of the same type (typically computed from different datasets) into a common, pooled estimate.

The function pool.anyist is the method for the class "anylist". It is used when the objects to be pooled are given in a list x.

Each of the elements of the list x, and each of the subsequent arguments ... if provided, must be an object of the same class.

Value

An object of the same class as each of the entries in x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

anylist, pool.

Examples

   Keach <- anylapply(waterstriders, Kest, ratio=TRUE, correction="iso")
   K <- pool(Keach)

[Package spatstat.explore version 3.3-4 Index]