pool.envelope {spatstat.explore}	R Documentation

Pool Data from Several Envelopes

Description

Pool the simulation data from several simulation envelopes (objects of class "envelope") and compute a new envelope.

Usage

## S3 method for class 'envelope'
pool(..., savefuns=FALSE, savepatterns=FALSE)

Arguments

...
Objects of class "envelope".

savefuns
Logical flag indicating whether to save all the simulated function values.

savepatterns
Logical flag indicating whether to save all the simulated point patterns.

Details

The function pool is generic. This is the method for the class "envelope" of simulation envelopes. It is used to combine the simulation data from several simulation envelopes and to compute an envelope based on the combined data.

Each of the arguments ... must be an object of class "envelope". These envelopes must be compatible, in that they are envelopes for the same function, and were computed using the same options.

In normal use, each envelope object will have been created by running the command envelope with the argument savefuns=TRUE. This ensures that each object contains the simulated data (summary function values for the simulated point patterns) that were used to construct the envelope.

The simulated data are extracted from each object and combined. A new envelope is computed from the combined set of simulations.

Alternatively, if each envelope object was created by running envelope with VARIANCE=TRUE, then the saved functions are not required.

The sample means and sample variances from each envelope will be pooled. A new envelope is computed from the pooled mean and variance.

Warnings or errors will be issued if the envelope objects ... appear to be incompatible. Apart from these basic checks, the code is not smart enough to decide whether it is sensible to pool the data.

To modify the envelope parameters or the type of envelope that is computed, first pool the envelope data using pool.envelope, then use envelope.envelope to modify the envelope parameters.

Value

An object of class "envelope".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

envelope, envelope.envelope, pool, pool.fasp

Examples

   E1 <- envelope(cells, Kest, nsim=10, savefuns=TRUE)
   E2 <- envelope(cells, Kest, nsim=20, savefuns=TRUE)
   pool(E1, E2)

   V1 <- envelope(E1, VARIANCE=TRUE)
   V2 <- envelope(E2, VARIANCE=TRUE)
   pool(V1, V2)

[Package spatstat.explore version 3.3-4 Index]