pool.fasp {spatstat.explore}	R Documentation

Pool Data from Several Function Arrays

Description

Pool the simulation data from several function arrays (objects of class "fasp") and compute a new function array.

Usage

## S3 method for class 'fasp'
pool(...)

Arguments

...
Objects of class "fasp".

Details

The function pool is generic. This is the method for the class "fasp" of function arrays. It is used to combine the simulation data from several arrays of simulation envelopes and to compute a new array of envelopes based on the combined data.

Each of the arguments ... must be a function array (object of class "fasp") containing simulation envelopes. This is typically created by running the command alltypes with the arguments envelope=TRUE and savefuns=TRUE. This ensures that each object is an array of simulation envelopes, and that each envelope contains the simulated data (summary function values) that were used to construct the envelope.

The simulated data are extracted from each object and combined. A new array of envelopes is computed from the combined set of simulations.

Warnings or errors will be issued if the objects ... appear to be incompatible. However, the code is not smart enough to decide whether it is sensible to pool the data.

Value

An object of class "fasp".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

fasp, alltypes, pool.envelope, pool

Examples

   A1 <- alltypes(amacrine,"K",nsim=9,envelope=TRUE,savefuns=TRUE)
   A2 <- alltypes(amacrine,"K",nsim=10,envelope=TRUE,savefuns=TRUE)
   pool(A1, A2)

[Package spatstat.explore version 3.3-4 Index]