pool.fv {spatstat.explore}	R Documentation

Pool Several Functions

Description

Combine several summary functions into a single function.

Usage

## S3 method for class 'fv'
pool(..., weights=NULL, relabel=TRUE, variance=TRUE)

Arguments

...
Objects of class "fv".

weights
Optional numeric vector of weights for the functions.

relabel
Logical value indicating whether the columns of the resulting function should be labelled to show that they were obtained by pooling.

variance
Logical value indicating whether to compute the sample variance and related terms.

Details

The function pool is generic. This is the method for the class "fv" of summary functions. It is used to combine several estimates of the same function into a single function.

Each of the arguments ... must be an object of class "fv". They must be compatible, in that they are estimates of the same function, and were computed using the same options.

The sample mean and sample variance of the corresponding estimates will be computed.

Value

An object of class "fv".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

pool, pool.anylist, pool.rat

Examples

   K <- lapply(waterstriders, Kest, correction="iso")
   Kall <- pool(K[[1]], K[[2]], K[[3]])
   Kall <- pool(as.anylist(K))
   plot(Kall, cbind(pooliso, pooltheo) ~ r,
              shade=c("loiso", "hiiso"),
              main="Pooled K function of waterstriders")

[Package spatstat.explore version 3.3-4 Index]