pool.quadrattest {spatstat.explore}	R Documentation

Pool Several Quadrat Tests

Description

Pool several quadrat tests into a single quadrat test.

Usage

## S3 method for class 'quadrattest'
pool(..., df=NULL, df.est=NULL, nsim=1999,
                                Xname=NULL, CR=NULL)

Arguments

...
Any number of objects, each of which is a quadrat test (object of class "quadrattest").

df
Optional. Number of degrees of freedom of the test statistic. Relevant only for chi^2 tests. Incompatible with df.est.

df.est
Optional. The number of fitted parameters, or the number of degrees of freedom lost by estimation of parameters. Relevant only for chi^2 tests. Incompatible with df.

nsim
Number of simulations, for Monte Carlo test.

Xname
Optional. Name of the original data.

CR
Optional. Numeric value of the Cressie-Read exponent CR overriding the value used in the tests.

Details

The function pool is generic. This is the method for the class "quadrattest".

An object of class "quadrattest" represents a chi^2 test or Monte Carlo test of goodness-of-fit for a point process model, based on quadrat counts. Such objects are created by the command quadrat.test.

Each of the arguments ... must be an object of class "quadrattest". They must all be the same type of test (chi-squared test or Monte Carlo test, conditional or unconditional) and must all have the same type of alternative hypothesis.

The test statistic of the pooled test is the Pearson X^2 statistic taken over all cells (quadrats) of all tests. The p value of the pooled test is then computed using either a Monte Carlo test or a chi^2 test.

For a pooled chi^2 test, the number of degrees of freedom of the combined test is computed by adding the degrees of freedom of all the tests (equivalent to assuming the tests are independent) unless it is determined by the arguments df or df.est. The resulting p value is computed to obtain the pooled test.

For a pooled Monte Carlo test, new simulations are performed to determine the pooled Monte Carlo p value.

Value

Another object of class "quadrattest".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

pool, quadrat.test

Examples

  Y <- split(humberside)
  test1 <- quadrat.test(Y[[1]])
  test2 <- quadrat.test(Y[[2]])
  pool(test1, test2, Xname="Humberside")

[Package spatstat.explore version 3.3-4 Index]