quadrat.test.splitppp {spatstat.explore}	R Documentation

Dispersion Test of CSR for Split Point Pattern Based on Quadrat Counts

Description

Performs a test of Complete Spatial Randomness for each of the component patterns in a split point pattern, based on quadrat counts. By default performs chi-squared tests; can also perform Monte Carlo based tests.

Usage

## S3 method for class 'splitppp'
quadrat.test(X, ..., df=NULL, df.est=NULL, Xname=NULL)

Arguments

X
A split point pattern (object of class "splitppp"), each component of which will be subjected to the goodness-of-fit test.

...
Arguments passed to quadrat.test.ppp.

df,df.est,Xname
Arguments passed to pool.quadrattest.

Details

The function quadrat.test is generic, with methods for point patterns (class "ppp"), split point patterns (class "splitppp") and point process models (class "ppm").

If X is a split point pattern, then for each of the component point patterns (taken separately) we test the null hypotheses of Complete Spatial Randomness, then combine the result into a single test.

The method quadrat.test.ppp is applied to each component point pattern. Then the results are pooled using pool.quadrattest to obtain a single test.

Value

An object of class "quadrattest" which can be printed and plotted.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

quadrat.test, quadratcount, quadrats, quadratresample, chisq.test, cdf.test.

To test a Poisson point process model against a specific Poisson alternative, use anova.ppm.

Examples

 qH <- quadrat.test(split(humberside), 2, 3)
 plot(qH)
 qH

[Package spatstat.explore version 3.3-4 Index]