radcumint {spatstat.explore}	R Documentation

Radial Cumulative Integral

Description

Compute the cumulative integral of an image over increasing radial distances from the origin.

Usage

radcumint(X, ..., origin, Xname, result = c("fv", "im"))

Arguments

X
A pixel image (object of class "im") with numerical or logical values.

...
Ignored.

origin
Optional. Origin about which the rotations should be performed. Either a numeric vector or a character string as described in the help for shift.owin.

Xname
Optional name for X to be used in the function labels.

result
Character string specifying the kind of result required: either a function object or a pixel image.

Details

This command computes, for each possible distance r, the integral of the pixel values lying inside the disc of radius r centred at the origin.

If result="fv" (the default) the result is a function object f of class "fv". For each value of radius r, the function value f(r) is the integral of X over the disc of radius r.

If result="im" the result is a pixel image, with the same dimensions as X. At a given pixel, the result is equal to f(r) where r is the distance from the given pixel to the origin. That is, at any given pixel, the resulting value is the integral of X over the disc centred at the origin whose boundary passes through the given pixel.

Value

An object of class "fv" or "im", with the same coordinate units as X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

rotmean, spatialcdf

Examples

   D <- density(redwood)
   plot(radcumint(D))
   plot(radcumint(D, result="im"))

[Package spatstat.explore version 3.3-4 Index]