rectcontact {spatstat.explore}	R Documentation

Contact Distribution Function using Rectangular Structuring Element

Description

Computes an estimate of the contact distribution function of a set, using a rectangular structuring element.

Usage

rectcontact(X, ..., asp = 1, npasses=4,
            eps = NULL, r = NULL, breaks = NULL, correction = c("rs", "km"))

Arguments

X
Logical-valued image. The TRUE values in the image determine the spatial region whose contact distribution function should be estimated.

...
Ignored.

asp
Aspect ratio for the rectangular metric. A single positive number. See rectdistmap for explanation.

npasses
Number of passes to perform in the distance algorithm. A positive integer. See rectdistmap for explanation.

eps
Pixel size, if the image should be converted to a finer grid.

r
Optional vector of distance values. Do Not Use This.

breaks
Do Not Use This.

correction
Character vector specifying the edge correction.

Details

To be written.

Value

Object of class "fv".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

Hest

Examples

  ## make an image which is TRUE/FALSE inside/outside the letter R
  V <- letterR
  Frame(V) <- grow.rectangle(Frame(V), 0.5)
  Z <- as.im(V, value=TRUE, na.replace=FALSE)
  ## analyse
  plot(rectcontact(Z))

[Package spatstat.explore version 3.3-4 Index]