relrisk {spatstat.explore}	R Documentation

Estimate of Spatially-Varying Relative Risk

Description

Generic command to estimate the spatially-varying probability of each type of point, or the ratios of such probabilities.

Usage

  relrisk(X, ...) 

Arguments

X
Either a point pattern (class "ppp") or a fitted point process model (class "ppm") from which the probabilities will be estimated.

...
Additional arguments appropriate to the method.

Details

In a point pattern containing several different types of points, we may be interested in the spatially-varying probability of each possible type, or the relative risks which are the ratios of such probabilities.

The command relrisk is generic and can be used to estimate relative risk in different ways.

The function relrisk.ppp is the method for point pattern datasets. It computes nonparametric estimates of relative risk by kernel smoothing.

The function relrisk.ppm is the method for fitted point process models (class "ppm"). It computes parametric estimates of relative risk, using the fitted model.

Value

A pixel image, or a list of pixel images, or a numeric vector or matrix, containing the requested estimates of relative risk.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

relrisk.ppp, relrisk.ppm.

[Package spatstat.explore version 3.3-4 Index]