relriskHeat {spatstat.explore}	R Documentation

Diffusion Estimate of Conditional Probabilities

Description

Computes the conditional probability estimator of relative risk based on a multitype point pattern using the diffusion estimate of the type-specific intensities.

Usage

relriskHeat(X, ...)

## S3 method for class 'ppp'
relriskHeat(X, ..., sigmaX=NULL, weights=NULL)

Arguments

X
A multitype point pattern (object of class "ppp").

...
Arguments passed to densityHeat controlling the estimation of each marginal intensity.

sigmaX
Optional. Numeric vector of bandwidths, one associated with each data point in X.

weights
Optional numeric vector of weights associated with each point of X.

Details

The function relriskHeat is generic. This file documents the method relriskHeat.ppp for spatial point patterns (objects of class "ppp").

This function estimates the spatially-varying conditional probability that a random point (given that it is present) will belong to a given type.

The algorithm separates X into the sub-patterns consisting of points of each type. It then applies densityHeat to each sub-pattern, using the same bandwidth and smoothing regimen for each sub-pattern, as specified by the arguments ....

If weights is specified, it should be a numeric vector of length equal to the number of points in X, so that weights[i] is the weight for data point X[i].

Similarly when performing lagged-arrival smoothing, the argument sigmaX must be a numeric vector of the same length as the number of points in X, and thus contain the point-specific bandwidths in the order corresponding to each of these points regardless of mark.

Value

A named list (of class solist) containing pixel images, giving the estimated conditional probability surfaces for each type.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Tilman Davies Tilman.Davies@otago.ac.nz.

References

Agarwal, N. and Aluru, N.R. (2010) A data-driven stochastic collocation approach for uncertainty quantification in MEMS. International Journal for Numerical Methods in Engineering 83, 575–597.

Baddeley, A., Davies, T., Rakshit, S., Nair, G. and McSwiggan, G. (2022) Diffusion smoothing for spatial point patterns. Statistical Science 37, 123–142.

Barry, R.P. and McIntyre, J. (2011) Estimating animal densities and home range in regions with irregular boundaries and holes: a lattice-based alternative to the kernel density estimator. Ecological Modelling 222, 1666–1672.

Botev, Z.I. and Grotowski, J.F. and Kroese, D.P. (2010) Kernel density estimation via diffusion. Annals of Statistics 38, 2916–2957.

See Also

relrisk.ppp for the traditional convolution-based kernel estimator of conditional probability surfaces, and the function risk in the sparr package for the density-ratio-based estimator.

Examples

  ## bovine tuberculosis data
  X <- subset(btb, select=spoligotype)
  plot(X) 
  P <- relriskHeat(X,sigma=9)
  plot(P)

[Package spatstat.explore version 3.3-4 Index]