rho2hat {spatstat.explore}	R Documentation

Smoothed Relative Density of Pairs of Covariate Values

Description

Given a point pattern and two spatial covariates Z1 and Z2, construct a smooth estimate of the relative risk of the pair (Z1, Z2).

Usage

rho2hat(object, cov1, cov2, ..., method=c("ratio", "reweight"))

Arguments

object
A point pattern (object of class "ppp"), a quadrature scheme (object of class "quad") or a fitted point process model (object of class "ppm").

cov1,cov2
The two covariates. Each argument is either a function(x,y) or a pixel image (object of class "im") providing the values of the covariate at any location, or one of the strings "x" or "y" signifying the Cartesian coordinates.

...
Additional arguments passed to density.ppp to smooth the scatterplots.

method
Character string determining the smoothing method. See Details.

Details

This is a bivariate version of rhohat.

If object is a point pattern, this command produces a smoothed version of the scatterplot of the values of the covariates cov1 and cov2 observed at the points of the point pattern.

The covariates cov1,cov2 must have continuous values.

If object is a fitted point process model, suppose X is the original data point pattern to which the model was fitted. Then this command assumes X is a realisation of a Poisson point process with intensity function of the form

lambda(u) = rho(Z1(u), Z2(u)) * kappa(u)

where kappa(u) is the intensity of the fitted model object, and rho(z1, z2) is a function to be estimated. The algorithm computes a smooth estimate of the function rho.

The method determines how the density estimates will be combined to obtain an estimate of rho(z1, z2):

If method="ratio", then rho(z1,z2) is estimated by the ratio of two density estimates. The numerator is a (rescaled) density estimate obtained by smoothing the points (Z1(y[i]), Z2(y[i])) obtained by evaluating the two covariate Z1, Z2 at the data points y[i]. The denominator is a density estimate of the reference distribution of (Z1, Z2).

If method="reweight", then rho(z1,z2) is estimated by applying density estimation to the points (Z1(y[i]), Z2(y[i])) obtained by evaluating the two covariate Z1, Z2 at the data points y[i], with weights inversely proportional to the reference density of (Z1, Z2).

Value

A pixel image (object of class "im"). Also belongs to the special class "rho2hat" which has a plot method.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Baddeley, A., Chang, Y.-M., Song, Y. and Turner, R. (2012) Nonparametric estimation of the dependence of a point process on spatial covariates. Statistics and Its Interface 5 (2), 221–236.

See Also

rhohat, methods.rho2hat

Examples

  attach(bei.extra)
  plot(rho2hat(bei, elev, grad))
  if(require("spatstat.model")) {
    fit <- ppm(bei ~elev, covariates=bei.extra)
    
      plot(rho2hat(fit, elev, grad))
    
    plot(rho2hat(fit, elev, grad, method="reweight"))
  }

[Package spatstat.explore version 3.3-4 Index]