rotmean {spatstat.explore}	R Documentation

Rotational Average of a Pixel Image

Description

Compute the average pixel value over all rotations of the image about the origin, as a function of distance from the origin.

Usage

rotmean(X, ..., origin, padzero=TRUE, Xname, result=c("fv", "im"), adjust=1)

Arguments

X
A pixel image.

...
Ignored.

origin
Optional. Origin about which the rotations should be performed. Either a numeric vector or a character string as described in the help for shift.owin.

padzero
Logical. If TRUE (the default), the value of X is assumed to be zero outside the window of X. If FALSE, the value of X is taken to be undefined outside the window of X.

Xname
Optional name for X to be used in the function labels.

result
Character string specifying the kind of result required: either a function object or a pixel image.

adjust
Adjustment factor for bandwidth used in kernel smoothing.

Details

This command computes, for each possible distance r, the average pixel value of the pixels lying at distance r from the origin. Kernel smoothing is used to obtain a smooth function of r.

If result="fv" (the default) the result is a function object of class "fv" giving the mean pixel value of X as a function of distance from the origin.

If result="im" the result is a pixel image, with the same dimensions as X, giving the mean value of X over all pixels lying at the same distance from the origin as the current pixel.

If padzero=TRUE (the default), the value of X is assumed to be zero outside the window of X. The rotational mean at a given distance r is the average value of the image X over the entire circle of radius r, including zero values outside the window if the circle lies partly outside the window.

If padzero=FALSE, the value of X is taken to be undefined outside the window of X. The rotational mean is the average of the X values over the subset of the circle of radius r that lies entirely inside the window.

Value

An object of class "fv" or "im", with the same coordinate units as X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

radcumint

Examples

  online <- interactive()
  resolution <- if(online) 128 else 32
  Z <- setcov(square(1), dimyx=resolution)
  f <- rotmean(Z)
  if(online) {
    plot(rotmean(Z))
    plot(rotmean(Z, result="im"))
  }

[Package spatstat.explore version 3.3-4 Index]