sdrPredict {spatstat.explore}	R Documentation

Compute Predictors from Sufficient Dimension Reduction

Description

Given the result of a Sufficient Dimension Reduction method, compute the new predictors.

Usage

  sdrPredict(covariates, B)

Arguments

covariates
A list of pixel images (objects of class "im").

B
Either a matrix of coefficients for the covariates, or the result of a call to sdr.

Details

This function assumes that sdr has already been used to find a minimal set of predictors based on the covariates. The argument B should be either the result of sdr or the coefficient matrix returned as one of the results of sdr. The columns of this matrix define linear combinations of the covariates. This function evaluates those linear combinations, and returns a list of pixel images containing the new predictors.

Value

A list of pixel images (objects of class "im") with one entry for each column of B.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

sdr

Examples

   A <- sdr(bei, bei.extra)
   Y <- sdrPredict(bei.extra, A)
   Y

[Package spatstat.explore version 3.3-4 Index]