spatialcdf {spatstat.explore}	R Documentation

Spatial Cumulative Distribution Function

Description

Compute the spatial cumulative distribution function of a spatial covariate, optionally using spatially-varying weights.

Usage

spatialcdf(Z, weights = NULL, normalise = FALSE, ..., W = NULL, Zname = NULL)

Arguments

Z
Spatial covariate. A pixel image or a function(x,y,...)

weights
Spatial weighting for different locations. A pixel image, a function(x,y,...), a window, a constant value, or a fitted point process model (object of class "ppm" or "kppm").

normalise
Logical. Whether the weights should be normalised so that they sum to 1.

...
Arguments passed to as.mask to determine the pixel resolution, or extra arguments passed to Z if it is a function.

W
Optional window (object of class "owin") defining the spatial domain.

Zname
Optional character string for the name of the covariate Z used in plots.

Details

If weights is missing or NULL, it defaults to 1. The values of the covariate Z are computed on a grid of pixels. The weighted cumulative distribution function of Z values is computed, taking each value with weight equal to the pixel area. The resulting function F is such that F(t) is the area of the region of space where Z <= t.

If weights is a pixel image or a function, then the values of weights and of the covariate Z are computed on a grid of pixels. The weights are multiplied by the pixel area. Then the weighted empirical cumulative distribution function of Z values is computed using ewcdf. The resulting function F is such that F(t) is the total weight (or weighted area) of the region of space where Z <= t.

If weights is a fitted point process model, then it should be a Poisson process. The fitted intensity of the model, and the value of the covariate Z, are evaluated at the quadrature points used to fit the model. The weights are multiplied by the weights of the quadrature points. Then the weighted empirical cumulative distribution of Z values is computed using ewcdf. The resulting function F is such that F(t) is the expected number of points in the point process that will fall in the region of space where Z <= t.

If normalise=TRUE, the function is normalised so that its maximum value equals 1, so that it gives the cumulative fraction of weight or cumulative fraction of points.

The result can be printed, plotted, and used as a function.

Value

A cumulative distribution function object belonging to the classes "spatialcdf", "ewcdf", "ecdf" (only if normalise=TRUE) and "stepfun".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

ewcdf, cdf.test

Examples

   with(bei.extra, {
     plot(spatialcdf(grad))
     if(require("spatstat.model")) {  
       fit <- ppm(bei ~ elev)
       plot(spatialcdf(grad, predict(fit)))
       A <- spatialcdf(grad, fit)
       A(0.1)
     }
  })
  plot(spatialcdf("x", W=letterR))

[Package spatstat.explore version 3.3-4 Index]