ssf {spatstat.explore}	R Documentation

Spatially Sampled Function

Description

Create an object that represents a spatial function which has been evaluated or sampled at an irregular set of points.

Usage

  ssf(loc, val)

Arguments

loc
The spatial locations at which the function has been evaluated. A point pattern (object of class "ppp").

val
The function values at these locations. A numeric vector with one entry for each point of loc, or a data frame with one row for each point of loc.

Details

An object of class "ssf" represents a real-valued or vector-valued function that has been evaluated or sampled at an irregular set of points. An example would be a spatial covariate that has only been measured at certain locations.

An object of this class also inherits the class "ppp", and is essentially the same as a marked point pattern, except for the class membership which enables it to be handled in a different way.

There are methods for plot, print etc; see plot.ssf and methods.ssf.

Use unmark to extract only the point locations, and marks.ssf to extract only the function values.

Value

Object of class "ssf".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

plot.ssf, methods.ssf, Smooth.ssf, with.ssf, [.ssf.

Examples

  ssf(cells, nndist(cells, k=1:3))

[Package spatstat.explore version 3.3-4 Index]