stienen {spatstat.explore}	R Documentation

Stienen Diagram

Description

Draw the Stienen diagram of a point pattern, or compute the region covered by the Stienen diagram.

Usage

stienen(X, ..., bg = "grey", border = list(bg = NULL))
stienenSet(X, edge=TRUE)

Arguments

X
Point pattern (object of class "ppp").

...
Arguments passed to plot.ppp to control the plot.

bg
Fill colour for circles.

border
Either a list of arguments passed to plot.ppp to control the display of circles at the border of the diagram, or the value FALSE indicating that the border circles should not be plotted.

edge
Logical value indicating whether to include the circles at the border of the diagram.

Details

The Stienen diagram of a point pattern (Stienen, 1982) is formed by drawing a circle around each point of the pattern, with diameter equal to the nearest-neighbour distance for that point. These circles do not overlap. If two points are nearest neighbours of each other, then the corresponding circles touch.

stienenSet(X) computes the union of these circles and returns it as a window (object of class "owin").

stienen(X) generates a plot of the Stienen diagram of the point pattern X. By default, circles are shaded in grey if they lie inside the window of X, and are not shaded otherwise.

Value

The plotting function stienen returns NULL.

The return value of stienenSet is a window (object of class "owin").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Stienen, H. (1982) Die Vergroeberung von Karbiden in reinen Eisen-Kohlenstoff Staehlen. Dissertation, RWTH Aachen.

See Also

nndist, plot.ppp

Examples

  Y <- stienenSet(cells)
  stienen(redwood)
  stienen(redwood, border=list(bg=NULL, lwd=2, cols="red"))

[Package spatstat.explore version 3.3-4 Index]