studpermu.test {spatstat.explore}	R Documentation

Studentised Permutation Test

Description

Perform a studentised permutation test for a difference between groups of point patterns.

Usage

   studpermu.test(X, formula, summaryfunction = Kest,
       ..., rinterval = NULL, nperm = 999,
        use.Tbar = FALSE, minpoints = 20, rsteps = 128,
        r = NULL, arguments.in.data = FALSE)

Arguments

X
Data. Either a hyperframe or a list of lists of point patterns.

formula
Formula describing the grouping, when X is a hyperframe. The left side of the formula identifies which column of X contains the point patterns. The right side identifies the grouping factor. If the formula is missing, the grouping variable is taken to be the first column of X that contains a factor, and the point patterns are taken from the first column that contains point patterns.

summaryfunction
Summary function applicable to point patterns.

...
Additional arguments passed to summaryfunction.

rinterval
Interval of distance values r over which the summary function should be evaluated and over which the test statistic will be integrated. If NULL, the default range of the summary statistic is used (taking the intersection of these ranges over all patterns).

nperm
Number of random permutations for the test.

use.Tbar
Logical value indicating choice of test statistic. If TRUE, use the alternative test statistic, which is appropriate for summary functions with roughly constant variance, such as K(r)/r or L(r).

minpoints
Minimum permissible number of points in a point pattern for inclusion in the test calculation.

rsteps
Number of discretisation steps in the rinterval.

r
Optional vector of distance values as the argument for summaryfunction. Should not usually be given. There is a sensible default.

arguments.in.data
Logical. If TRUE, individual extra arguments to summaryfunction will be taken from X (which must be a hyperframe). This assumes that the first argument of summaryfunction is the point pattern dataset.

Details

This function performs the studentized permutation test of Hahn (2012) for a difference between groups of point patterns.

The first argument X should be either

a list of lists of point patterns.

Each element of X will be interpreted as a group of point patterns, assumed to be replicates of the same point process.

a hyperframe:

One column of the hyperframe should contain point patterns, and another column should contain a factor indicating the grouping. The argument formula should be a formula in the R language specifying the grouping: it should be of the form P ~ G where P is the name of the column of point patterns, and G is the name of the factor.

A group needs to contain at least two point patterns with at least minpoints points in each pattern.

The function returns an object of class "htest" and "studpermutest" that can be printed and plotted. The printout shows the test result and p-value. The plot shows the summary functions for the groups (and the group means if requested).

Value

Object of class "studpermutest".

Author(s)

Ute Hahn.

Modified for spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Hahn, U. (2012) A studentized permutation test for the comparison of spatial point patterns. Journal of the American Statistical Association 107 (498), 754–764.

See Also

plot.studpermutest

Examples

  np <- if(interactive()) 99 else 19
  testpyramidal <- studpermu.test(pyramidal, Neurons ~ group, nperm=np)
  testpyramidal

[Package spatstat.explore version 3.3-4 Index]