subspaceDistance {spatstat.explore}	R Documentation

Distance Between Linear Spaces

Description

Evaluate the distance between two linear subspaces using the measure proposed by Li, Zha and Chiaromonte (2005).

Usage

subspaceDistance(B0, B1)

Arguments

B0
Matrix whose columns are a basis for the first subspace.

B1
Matrix whose columns are a basis for the second subspace.

Details

This algorithm calculates the maximum absolute value of the eigenvalues of P1-P0 where P0,P1 are the projection matrices onto the subspaces generated by B0,B1. This measure of distance was proposed by Li, Zha and Chiaromonte (2005). See also Xia (2007).

Value

A single numeric value.

Author(s)

Matlab original by Yongtao Guan, translated to R by Suman Rakshit.

References

Guan, Y. and Wang, H. (2010) Sufficient dimension reduction for spatial point processes directed by Gaussian random fields. Journal of the Royal Statistical Society, Series B, 72, 367–387.

Li, B., Zha, H. and Chiaromonte, F. (2005) Contour regression: a general approach to dimension reduction. Annals of Statistics 33, 1580–1616.

Xia, Y. (2007) A constructive approach to the estimation of dimension reduction directions. Annals of Statistics 35, 2654–2690.

[Package spatstat.explore version 3.3-4 Index]