with.ssf {spatstat.explore}	R Documentation

Evaluate Expression in a Spatially Sampled Function

Description

Given a spatially sampled function, evaluate an expression involving the function values.

Usage

  apply.ssf(X, ...)

  ## S3 method for class 'ssf'
with(data, ...)

Arguments

X, data
A spatially sampled function (object of class "ssf").

...
Arguments passed to with.default or apply specifying what to compute.

Details

An object of class "ssf" represents a function (real- or vector-valued) that has been sampled at a finite set of points. It contains a data frame which provides the function values at the sample points.

In with.ssf, the expression specified by ... will be evaluated in this dataframe. In apply.ssf, the dataframe will be subjected to the apply operator using the additional arguments ....

If the result of evaluation is a data frame with one row for each data point, or a numeric vector with one entry for each data point, then the result will be an object of class "ssf" containing this information. Otherwise, the result will be a numeric vector.

Value

An object of class "ssf" or a numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

ssf

Examples

  a <- ssf(cells, data.frame(d=nndist(cells), i=1:npoints(cells)))
  with(a, i/d)

[Package spatstat.explore version 3.3-4 Index]