Gdot.inhom {spatstat.explore}	R Documentation

Inhomogeneous Multitype G Dot Function

Description

For a multitype point pattern, estimate the inhomogeneous version of the dot G function, which is the distribution of the distance from a point of type i to the nearest other point of any type, adjusted for spatially varying intensity.

Usage

   Gdot.inhom(X, i,
              lambdaI = NULL, lambdadot = NULL, lambdamin = NULL,
              ...,
              r = NULL, ReferenceMeasureMarkSetI = NULL, ratio = FALSE)

Arguments

X
The observed point pattern, from which an estimate of the inhomogeneous dot type G function Gi.(r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor). See under Details.

i
The type (mark value) of the points in X from which distances are measured. A character string (or something that will be converted to a character string). Defaults to the first level of marks(X).

lambdaI
Optional. Values of the estimated intensity of the sub-process of points of type i. Either a pixel image (object of class "im"), a numeric vector containing the intensity values at each of the type i points in X, a fitted point process model (object of class "ppm" or "kppm" or "dppm"), or a function(x,y) which can be evaluated to give the intensity value at any location.

lambdadot
Optional. Values of the estimated intensity of the entire point process, Either a pixel image (object of class "im"), a numeric vector containing the intensity values at each of the points in X, a fitted point process model (object of class "ppm" or "kppm" or "dppm"), or a function(x,y) which can be evaluated to give the intensity value at any location.

lambdamin
Optional. The minimum possible value of the intensity over the spatial domain. A positive numerical value.

...
Ignored.

r
vector of values for the argument r at which the inhomogeneous dot type G function Gi.(r) should be evaluated. Not normally given by the user; there is a sensible default.

ReferenceMeasureMarkSetI
Optional. The total measure of the mark set. A positive number.

ratio
Logical value indicating whether to save ratio information.

Details

This is a generalisation of the function Gdot to include an adjustment for spatially inhomogeneous intensity, in a manner similar to the function Ginhom.

The argument lambdaI supplies the values of the intensity of the sub-process of points of type i. It may be either

a pixel image

(object of class "im") which gives the values of the type i intensity at all locations in the window containing X;

a numeric vector

containing the values of the type i intensity evaluated only at the data points of type i. The length of this vector must equal the number of type i points in X.

a function

of the form function(x,y) which can be evaluated to give values of the intensity at any locations.

a fitted point process model

(object of class "ppm", "kppm" or "dppm") whose fitted trend can be used as the fitted intensity. (If update=TRUE the model will first be refitted to the data X before the trend is computed.)

omitted:

if lambdaI is omitted then it will be estimated using a leave-one-out kernel smoother.

If lambdaI is omitted, then it will be estimated using a ‘leave-one-out’ kernel smoother.

Similarly the argument lambdadot should contain estimated values of the intensity of the entire point process. It may be either a pixel image, a numeric vector of length equal to the number of points in X, a function, or omitted.

The argument r is the vector of values for the distance r at which Gi.(r) should be evaluated. The values of r must be increasing nonnegative numbers and the maximum r value must not exceed the radius of the largest disc contained in the window.

Value

An object of class "fv" (see fv.object) containing estimates of the inhomogeneous dot type G function.

Warnings

The argument i is interpreted as a level of the factor X$marks. It is converted to a character string if it is not already a character string. The value i=1 does not refer to the first level of the factor.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Cronie, O. and Van Lieshout, M.N.M. (2015) Summary statistics for inhomogeneous marked point processes. Annals of the Institute of Statistical Mathematics DOI: 10.1007/s10463-015-0515-z

See Also

Gdot, Ginhom, Gcross.inhom, Gmulti.inhom.

Examples

  X <- rescale(amacrine)
  if(interactive() && require(spatstat.model)) {
    ## how to do it normally
    mod <- ppm(X ~ marks * x)
    lam <- fitted(mod, dataonly=TRUE)
    lmin <- min(predict(mod)[["off"]]) * 0.9
  } else {
    ## for package testing 
    lam <- intensity(X)[as.integer(marks(X))]
    lmin <- intensity(X)[2] * 0.9
  }
  lamI <- lam[marks(X) == "on"]
  GD <- Gdot.inhom(X, "on", lambdaI=lamI, lambdadot=lam, lambdamin=lmin)

[Package spatstat.explore version 3.3-4 Index]