SmoothHeat {spatstat.explore}	R Documentation

Spatial Smoothing of Data by Diffusion

Description

Generic function to perform spatial smoothing of spatial data by diffusion.

Usage

  SmoothHeat(X, ...)

Arguments

X
Some kind of spatial data

...
Arguments passed to methods.

Details

This generic function calls an appropriate method to perform spatial smoothing on the spatial dataset X using diffusion.

Methods for this function include

SmoothHeat.ppp for point patterns

SmoothHeat.im for pixel images.

Value

An object containing smoothed values of the input data, in an appropriate format. See the documentation for the methods.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

SmoothHeat.ppp, SmoothHeat.im.

[Package spatstat.explore version 3.3-4 Index]