SmoothHeat.ppp {spatstat.explore}	R Documentation

Spatial Smoothing of Observations using Diffusion Estimate of Density

Description

Performs spatial smoothing of numeric values observed at a set of irregular locations, using the diffusion estimate of the density.

Usage

## S3 method for class 'ppp'
SmoothHeat(X, sigma, ..., weights=NULL)

Arguments

X
Point pattern (object of class "ppp") with numeric marks.

sigma
Smoothing bandwidth. A single number giving the equivalent standard deviation of the smoother.

...
Arguments passed to densityHeat controlling the estimation of each marginal intensity, or passed to pixellate.ppp controlling the pixel resolution.

weights
Optional numeric vector of weights associated with each data point.

Details

This is the analogue of the Nadaraya-Watson smoother, using the diffusion smoothing estimation procedure (Baddeley et al, 2022). The numerator and denominator of the Nadaraya-Watson smoother are calculated using densityHeat.ppp.

Value

Pixel image (object of class "im") giving the smoothed mark value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Tilman Davies Tilman.Davies@otago.ac.nz and Suman Rakshit.

References

Baddeley, A., Davies, T., Rakshit, S., Nair, G. and McSwiggan, G. (2022) Diffusion smoothing for spatial point patterns. Statistical Science 37, 123–142.

See Also

Smooth.ppp for the usual kernel-based smoother (the Nadaraya-Watson smoother) and densityHeat for the diffusion estimate of density.

Examples

   plot(SmoothHeat(longleaf, 10))

[Package spatstat.explore version 3.3-4 Index]