bw.stoyan {spatstat.explore}	R Documentation

Stoyan's Rule of Thumb for Bandwidth Selection

Description

Computes a rough estimate of the appropriate bandwidth for kernel smoothing estimators of the pair correlation function and other quantities.

Usage

bw.stoyan(X, co=0.15)

Arguments

X
A point pattern (object of class "ppp").

co
Coefficient appearing in the rule of thumb. See Details.

Details

Estimation of the pair correlation function and other quantities by smoothing methods requires a choice of the smoothing bandwidth. Stoyan and Stoyan (1995, equation (15.16), page 285) proposed a rule of thumb for choosing the smoothing bandwidth.

For the Epanechnikov kernel, the rule of thumb is to set the kernel's half-width h to 0.15/sqrt(λ) where λ is the estimated intensity of the point pattern, typically computed as the number of points of X divided by the area of the window containing X.

For a general kernel, the corresponding rule is to set the standard deviation of the kernel to σ = 0.15/sqrt(5 * λ).

The coefficient 0.15 can be tweaked using the argument co.

To ensure the bandwidth is finite, an empty point pattern is treated as if it contained 1 point.

Value

A finite positive numerical value giving the selected bandwidth (the standard deviation of the smoothing kernel).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Stoyan, D. and Stoyan, H. (1995) Fractals, random shapes and point fields: methods of geometrical statistics. John Wiley and Sons.

See Also

pcf, bw.relrisk

Examples

  bw.stoyan(shapley)

[Package spatstat.explore version 3.3-4 Index]