circdensity {spatstat.explore}	R Documentation

Density Estimation for Circular Data

Description

Computes a kernel smoothed estimate of the probability density for angular data.

Usage

circdensity(x, sigma = "nrd0", ...,
               bw = NULL,
               weights=NULL, unit = c("degree", "radian"))

Arguments

x
Numeric vector, containing angular data.

sigma
Smoothing bandwidth, or bandwidth selection rule, passed to density.default.

bw
Alternative to sigma for consistency with other functions.

...
Additional arguments passed to density.default, such as kernel and weights.

weights
Optional numeric vector of weights for the data in x.

unit
The unit of angle in which x is expressed.

Details

The angular values x are smoothed using (by default) the wrapped Gaussian kernel with standard deviation sigma.

Value

An object of class "density" (produced by density.default) which can be plotted by plot or by rose.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

density.default), rose.

Examples

  ang <- runif(1000, max=360)
  rose(circdensity(ang, 12))

[Package spatstat.explore version 3.3-4 Index]