clarkevans {spatstat.explore}	R Documentation

Clark and Evans Aggregation Index

Description

Computes the Clark and Evans aggregation index R for a spatial point pattern.

Usage

clarkevans(X, correction=c("none", "Donnelly", "cdf"),
              clipregion=NULL)

Arguments

X
A spatial point pattern (object of class "ppp").

correction
Character vector. The type of edge correction(s) to be applied.

clipregion
Clipping region for the guard area correction. A window (object of class "owin"). See Details.

Details

The Clark and Evans (1954) aggregation index R is a crude measure of clustering or ordering of a point pattern. It is the ratio of the observed mean nearest neighbour distance in the pattern to that expected for a Poisson point process of the same intensity. A value R>1 suggests ordering, while R<1 suggests clustering.

Without correction for edge effects, the value of R will be positively biased. Edge effects arise because, for a point of X close to the edge of the window, the true nearest neighbour may actually lie outside the window. Hence observed nearest neighbour distances tend to be larger than the true nearest neighbour distances.

The argument correction specifies an edge correction or several edge corrections to be applied. It is a character vector containing one or more of the options "none", "Donnelly", "guard" and "cdf" (which are recognised by partial matching). These edge corrections are:

"none":

No edge correction is applied.

"Donnelly":

Edge correction of Donnelly (1978), available for rectangular windows only. The theoretical expected value of mean nearest neighbour distance under a Poisson process is adjusted for edge effects by the edge correction of Donnelly (1978). The value of R is the ratio of the observed mean nearest neighbour distance to this adjusted theoretical mean.

"guard":

Guard region or buffer area method. The observed mean nearest neighbour distance for the point pattern X is re-defined by averaging only over those points of X that fall inside the sub-window clipregion.

"cdf":

Cumulative Distribution Function method. The nearest neighbour distance distribution function G(r) of the stationary point process is estimated by Gest using the Kaplan-Meier type edge correction. Then the mean of the distribution is calculated from the cdf.

Alternatively correction="all" selects all options.

If the argument clipregion is given, then the selected edge corrections will be assumed to include correction="guard".

To perform a test based on the Clark-Evans index, see clarkevans.test.

Value

A numeric value, or a numeric vector with named components

naive
R without edge correction

Donnelly
R using Donnelly edge correction

guard
R using guard region

cdf
R using cdf method

(as selected by correction). The value of the Donnelly component will be NA if the window of X is not a rectangle.

Author(s)

John Rudge rudge@esc.cam.ac.uk with modifications by Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Clark, P.J. and Evans, F.C. (1954) Distance to nearest neighbour as a measure of spatial relationships in populations Ecology 35, 445–453.

Donnelly, K. (1978) Simulations to determine the variance and edge-effect of total nearest neighbour distance. In I. Hodder (ed.) Simulation studies in archaeology, Cambridge/New York: Cambridge University Press, pp 91–95.

See Also

clarkevans.test, hopskel, nndist, Gest

Examples

  # Example of a clustered pattern
  clarkevans(redwood)

  # Example of an ordered pattern
  clarkevans(cells)

  # Random pattern
  X <- rpoispp(100)
  clarkevans(X)

  # How to specify a clipping region
  clip1 <- owin(c(0.1,0.9),c(0.1,0.9))
  clip2 <- erosion(Window(cells), 0.1)
  clarkevans(cells, clipregion=clip1)
  clarkevans(cells, clipregion=clip2)

[Package spatstat.explore version 3.3-4 Index]