markmarkscatter {spatstat.explore}	R Documentation

Mark-Mark Scatter Plot

Description

Generates the mark-mark scatter plot of a point pattern.

Usage

markmarkscatter(X, rmax, ..., col = NULL, symap = NULL, transform=I, jit=FALSE)

Arguments

X
A point pattern (object of class "ppp", "pp3", "lpp" or "ppx") with numeric marks.

rmax
Maximum distance between pairs of points which contribute to the plot.

...
Additional arguments passed to plot.ppp to control the scatterplot.

transform
Optional. A function which should be applied to the mark values.

jit
Logical value indicating whether mark values should be randomly perturbed using jitter.

col
Optional. A vector of colour values, or a colourmap to be used to portray the pairwise distance values. Ignored if symap is given.

symap
Optional. A symbolmap to be used to portray the pairwise distance values. Overrides col.

Details

The mark-mark scatter plot (Ballani et al, 2019) is a scatterplot of the mark values of all pairs of distinct points in X which are closer than the distance rmax. The dots in the scatterplot are coloured according to the pairwise distance between the two spatial points. The plot is augmented by three curves explained by Ballani et al (2019).

If the marks only take a few different values, then it is usually appropriate to apply random perturbation (jitter) to the mark values, by setting jit=TRUE.

Value

Null.

Author(s)

Adrian Baddeley (coded from the description in Ballani et al.)

References

Ballani, F., Pommerening, A. and Stoyan, D. (2019) Mark-mark scatterplots improve pattern analysis in spatial plant ecology. Ecological Informatics 49, 13–21.

Examples

  markmarkscatter(longleaf, 10)

  markmarkscatter(spruces, 10, jit=TRUE)

[Package spatstat.explore version 3.3-4 Index]