pcfcross.inhom {spatstat.explore}	R Documentation

Inhomogeneous Multitype Pair Correlation Function (Cross-Type)

Description

Estimates the inhomogeneous cross-type pair correlation function for a multitype point pattern.

Usage

pcfcross.inhom(X, i, j, lambdaI = NULL, lambdaJ = NULL, ...,
               r = NULL, breaks = NULL,
               kernel="epanechnikov", bw=NULL, adjust.bw = 1, stoyan=0.15,
               correction = c("isotropic", "Ripley", "translate"),
               sigma = NULL, adjust.sigma = 1, varcov = NULL)

Arguments

X
The observed point pattern, from which an estimate of the inhomogeneous cross-type pair correlation function g[i,j](r) will be computed. It must be a multitype point pattern (a marked point pattern whose marks are a factor).

i
The type (mark value) of the points in X from which distances are measured. A character string (or something that will be converted to a character string). Defaults to the first level of marks(X).

j
The type (mark value) of the points in X to which distances are measured. A character string (or something that will be converted to a character string). Defaults to the second level of marks(X).

lambdaI
Optional. Values of the estimated intensity function of the points of type i. Either a vector giving the intensity values at the points of type i, a pixel image (object of class "im") giving the intensity values at all locations, or a function(x,y) which can be evaluated to give the intensity value at any location.

lambdaJ
Optional. Values of the estimated intensity function of the points of type j. A numeric vector, pixel image or function(x,y).

r
Vector of values for the argument r at which g[i,j](r) should be evaluated. There is a sensible default.

breaks
This argument is for internal use only.

kernel
Choice of one-dimensional smoothing kernel, passed to density.default.

bw
Bandwidth for one-dimensional smoothing kernel, passed to density.default.

adjust.bw
Numeric value. bw will be multiplied by this value.

...
Other arguments passed to the one-dimensional kernel density estimation function density.default.

stoyan
Bandwidth coefficient; see Details.

correction
Choice of edge correction.

sigma,varcov
Optional arguments passed to density.ppp to control the smoothing bandwidth, when lambdaI or lambdaJ is estimated by spatial kernel smoothing.

adjust.sigma
Numeric value. sigma will be multiplied by this value.

Details

The inhomogeneous cross-type pair correlation function g[i,j](r) is a summary of the dependence between two types of points in a multitype spatial point process that does not have a uniform density of points.

The best intuitive interpretation is the following: the probability p(r) of finding two points, of types i and j respectively, at locations x and y separated by a distance r is equal to

p(r) = lambda[i](x) * lambda[j](y) * g(r) dx dy

where lambda[i] is the intensity function of the process of points of type i. For a multitype Poisson point process, this probability is p(r) = lambda[i](x) * lambda[j](y) so g[i,j](r) = 1.

The command pcfcross.inhom estimates the inhomogeneous pair correlation using a modified version of the algorithm in pcf.ppp. The arguments bw and adjust.bw control the degree of one-dimensional smoothing of the estimate of pair correlation.

If the arguments lambdaI and/or lambdaJ are missing or null, they will be estimated from X by spatial kernel smoothing using a leave-one-out estimator, computed by density.ppp. The arguments sigma, varcov and adjust.sigma control the degree of spatial smoothing.

Value

A function value table (object of class "fv"). Essentially a data frame containing the variables

r
the vector of values of the argument r at which the inhomogeneous cross-type pair correlation function g[i,j](r) has been estimated

theo
vector of values equal to 1, the theoretical value of g[i,j](r) for the Poisson process

trans
vector of values of g[i,j](r) estimated by translation correction

iso
vector of values of g[i,j](r) estimated by Ripley isotropic correction

as required.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

pcf.ppp, pcfinhom, pcfcross, pcfdot.inhom

Examples

  plot(pcfcross.inhom(amacrine, "on", "off", stoyan=0.1),
       legendpos="bottom")

[Package spatstat.explore version 3.3-4 Index]