sdr {spatstat.explore}	R Documentation

Sufficient Dimension Reduction

Description

Given a point pattern and a set of predictors, find a minimal set of new predictors, each constructed as a linear combination of the original predictors.

Usage

sdr(X, covariates, ...)

## S3 method for class 'ppp'
sdr(X, covariates,
                  method = c("DR", "NNIR", "SAVE", "SIR", "TSE"),
                  Dim1 = 1, Dim2 = 1, predict=FALSE, ...)

Arguments

X
A point pattern (object of class "ppp").

covariates
A list of pixel images (objects of class "im") to serve as predictor variables.

method
Character string indicating which method to use. See Details.

Dim1
Dimension of the first order Central Intensity Subspace (applicable when method is "DR", "NNIR", "SAVE" or "TSE").

Dim2
Dimension of the second order Central Intensity Subspace (applicable when method="TSE").

predict
Logical value indicating whether to compute the new predictors as well.

...
Additional arguments (ignored by sdr.ppp).

Details

Given a point pattern X and predictor variables Z[1], ..., Z[p], Sufficient Dimension Reduction methods (Guan and Wang, 2010) attempt to find a minimal set of new predictor variables, each constructed by taking a linear combination of the original predictors, which explain the dependence of X on Z[1], ..., Z[p]. The methods do not assume any particular form of dependence of the point pattern on the predictors. The predictors are assumed to be Gaussian random fields.

Available methods are:

method="DR" 	 directional regression

method="NNIR" 	 nearest neighbour inverse regression

method="SAVE" 	 sliced average variance estimation

method="SIR" 	 sliced inverse regression

method="TSE" 	 two-step estimation

The result includes a matrix B whose columns are estimates of the basis vectors of the space of new predictors. That is, the jth column of B expresses the jth new predictor as a linear combination of the original predictors.

If predict=TRUE, the new predictors are also evaluated. They can also be evaluated using sdrPredict.

Value

A list with components B, M or B, M1, M2 where B is a matrix whose columns are estimates of the basis vectors for the space, and M or M1,M2 are matrices containing estimates of the kernel.

If predict=TRUE, the result also includes a component Y which is a list of pixel images giving the values of the new predictors.

Author(s)

Matlab original by Yongtao Guan, translated to R by Suman Rakshit.

References

Guan, Y. and Wang, H. (2010) Sufficient dimension reduction for spatial point processes directed by Gaussian random fields. Journal of the Royal Statistical Society, Series B, 72, 367–387.

See Also

sdrPredict to compute the new predictors from the coefficient matrix.

dimhat to estimate the subspace dimension.

subspaceDistance

Examples

   A <- sdr(bei, bei.extra, predict=TRUE)
   A
   Y1 <- A$Y[[1]]
   plot(Y1)
   points(bei, pch=".", cex=2)
   # investigate likely form of dependence
   plot(rhohat(bei, Y1))

[Package spatstat.explore version 3.3-4 Index]