sharpen {spatstat.explore}	R Documentation

Data Sharpening of Point Pattern

Description

Performs Choi-Hall data sharpening of a spatial point pattern.

Usage

sharpen(X, ...)
## S3 method for class 'ppp'
sharpen(X, sigma=NULL, ...,
                      varcov=NULL, edgecorrect=FALSE)

Arguments

X
A marked point pattern (object of class "ppp").

sigma
Standard deviation of isotropic Gaussian smoothing kernel.

varcov
Variance-covariance matrix of anisotropic Gaussian kernel. Incompatible with sigma.

edgecorrect
Logical value indicating whether to apply edge effect bias correction.

...
Arguments passed to density.ppp to control the pixel resolution of the result.

Details

Choi and Hall (2001) proposed a procedure for data sharpening of spatial point patterns. This procedure is appropriate for earthquake epicentres and other point patterns which are believed to exhibit strong concentrations of points along a curve. Data sharpening causes such points to concentrate more tightly along the curve.

If the original data points are X[1],..., X[n] then the sharpened points are

X^[i] = (sum[j] X[j] * k(X[j] - X[i]))/(sum[j] k(X[j] - X[i]))

where k is a smoothing kernel in two dimensions. Thus, the new point X^[i] is a vector average of the nearby points X[j].

The function sharpen is generic. It currently has only one method, for two-dimensional point patterns (objects of class "ppp").

If sigma is given, the smoothing kernel is the isotropic two-dimensional Gaussian density with standard deviation sigma in each axis. If varcov is given, the smoothing kernel is the Gaussian density with variance-covariance matrix varcov.

The data sharpening procedure tends to cause the point pattern to contract away from the boundary of the window. That is, points X[i] that lie 'quite close to the edge of the window of the point pattern tend to be displaced inward. If edgecorrect=TRUE then the algorithm is modified to correct this vector bias.

Value

A point pattern (object of class "ppp") in the same window as the original pattern X, and with the same marks as X.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Choi, E. and Hall, P. (2001) Nonparametric analysis of earthquake point-process data. In M. de Gunst, C. Klaassen and A. van der Vaart (eds.) State of the art in probability and statistics: Festschrift for Willem R. van Zwet, Institute of Mathematical Statistics, Beachwood, Ohio. Pages 324–344.

See Also

density.ppp, Smooth.ppp.

Examples

   X <- unmark(shapley)
   
   Y <- sharpen(X, sigma=0.5)
   Z <- sharpen(X, sigma=0.5, edgecorrect=TRUE)
   opa <- par(mar=rep(0.2, 4))
   plot(solist(X, Y, Z), main= " ",
        main.panel=c("data", "sharpen", "sharpen, correct"),
        pch=".", equal.scales=TRUE, mar.panel=0.2)
   par(opa)

[Package spatstat.explore version 3.3-4 Index]